% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_base_functions.R
\name{add_time_col}
\alias{add_time_col}
\title{Add Time Column}
\usage{
add_time_col(df, time_value, time_id = "time")
}
\arguments{
\item{df}{The data frame to which the time column will be added.}

\item{time_value}{The value to be assigned to the time column.}

\item{time_id}{The name of the time column (default is "time").}
}
\value{
The data frame with the added time column.
}
\description{
This function adds a time column to a data frame.
}
\examples{
df <- data.frame(x = 1:5, y = letters[1:5])
add_time_col(df, "2022-01-01")

}
\seealso{
Other Post-calculation: 
\code{\link{impute_all}()},
\code{\link{post_calc_autojoin}()},
\code{\link{post_calc_convert_time}()},
\code{\link{post_calc_merge_all}()},
\code{\link{post_calc_merge_features}()},
\code{\link{post_calc_unify_timecols}()},
\code{\link{reduce_merge}()}
}
\concept{Post-calculation}
