% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_learner.R
\name{attach_pred}
\alias{attach_pred}
\title{Combine base learner prediction values with testing data.}
\usage{
attach_pred(
  pred = NULL,
  test = NULL,
  target_cols = c("site_id", "time", "lon", "lat"),
  yvar = "Arithmetic.Mean"
)
}
\arguments{
\item{pred}{data.table(1). Predicted values from \code{fit_prediction}}

\item{test}{data.table(1). \code{data.table} with with testing dataset.}

\item{target_cols}{characters(1). Columns to retain from the full
data.frame. DO NOT include \code{yvar} variable in the \code{target_cols}.}

\item{yvar}{character(1). Outcome variable name.}
}
\value{
a data.frame object, including the target columns from \code{data} and
the predictions for each base learner.
}
\description{
This function combines outcome data (observations) with the prediction
values from each base learner.
}
\keyword{Utility}
\keyword{meta_learner}
