% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_learner.R
\name{base_learner_cv_fit}
\alias{base_learner_cv_fit}
\title{Split training-test data for base learner fitting}
\usage{
base_learner_cv_fit(learner, ymat, xmat, cv_index, fun, ...)
}
\arguments{
\item{learner}{character(1). Currently one of 'randomforest', 'xgboost',
and 'cnn'}

\item{ymat}{data.frame or matrix. Dependent variable.}

\item{xmat}{data.frame or matrix. Independent variables.}

\item{cv_index}{integer. Index per cross-validation method.
See \code{?generate_cv_index} for details.}

\item{fun}{base_learner_fit_*}

\item{...}{Arguments passed to the argument \code{fun}}
}
\value{
List of 4 with xtrain, ytrain, xtest, and ytest.
}
\description{
Split training-test data for base learner fitting
}
\author{
Insang Song
}
