% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_learner.R
\name{base_learner_cv_outcome}
\alias{base_learner_cv_outcome}
\title{Return cvfold outcome values}
\usage{
base_learner_cv_outcome(data, cv_index, dependent_name)
}
\arguments{
\item{data}{stdt. See \code{\link{convert_stobj_to_stdt}}}

\item{cv_index}{integer. Index per cross-validation method.
See \code{\link{generate_cv_index}}}

\item{dependent_name}{character(1). Name of the dependent variable.}
}
\value{
List length of \code{max(cv_index)}
with numeric vectors when \code{length(dependent_name) == 1} or
data.frames \code{length(dependent_name) > 1}
}
\description{
Return cvfold outcome values
}
\author{
Insang Song
}
