% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_learner.R
\name{base_learner_fit}
\alias{base_learner_fit}
\title{Fit base learner}
\usage{
base_learner_fit(
  data,
  learner = c("randomforest", "xgboost", "cnn"),
  dependent_name = "pm2.5",
  independent_name = NULL,
  cv_mode = c("lolo", "loto", "lolto", "random", "lblo", "lbto", "lblto"),
  cv_fold = 5L,
  sp_fold = NULL,
  t_fold = NULL,
  blocks = NULL,
  block_id = NULL,
  return_full_object = FALSE
)
}
\arguments{
\item{data}{stdt. See \link{\code{convert_stobj_to_stdt}}}

\item{learner}{character(1). Currently one of 'randomforest', 'xgboost',
and 'cnn'}

\item{dependent_name}{character(1). Name of the dependent variable.}

\item{independent_name}{character(1). Names of independent variables.}

\item{cv_mode}{character(1). One of
\code{'lolo'} (leave-one-location-out),
\code{'loto'} (leave-one-time-out),
\code{'lolto'} (leave-one-location-time-out),
\code{'lblo'} (leave-block-location-out),
\code{'lbto'} (leave-block-time-out),
\code{'lblto'} (leave-block-location-time-out), and
\code{'random'} (full random selection)}

\item{cv_fold}{integer(1). Number of folds for cross-validation.}

\item{sp_fold}{integer(1). Number of subfolds for spatial blocks.}

\item{t_fold}{integer(1). Number of subfolds for temporal blocks.}

\item{blocks}{integer(2)/sf/SpatVector object.}

\item{block_id}{character(1). The unique identifier of each block.}

\item{return_full_object}{logical(1).
\code{TRUE} will return the object whose class is dependent on
the package in use, whereas \code{FALSE} will only return the prediction
with spatial/temporal information. Full object is potentially helpful for
debugging.}
}
\value{
data.frame
}
\description{
This function fits a selected base learner, which is supposed
to be one of random forest, XGBoost, and convolutional neural network,
using one of spatiotemporal cross-validation approaches.
}
\author{
Insang Song
}
