% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_learner.R
\name{base_learner_fit_ranger}
\alias{base_learner_fit_ranger}
\title{Fit random forests with ranger}
\usage{
base_learner_fit_ranger(ymat, xmat, ...)
}
\arguments{
\item{ymat}{data.frame or matrix. Dependent variable.}

\item{xmat}{data.frame or matrix. Independent variables.}

\item{...}{Arguments passed to \code{ranger::ranger}}
}
\value{
ranger object.
}
\description{
Fit random forests with ranger
}
\author{
Insang Song
}
