% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_learner.R
\name{base_learner_fit_xgboost}
\alias{base_learner_fit_xgboost}
\title{Fit XGBoost model}
\usage{
base_learner_fit_xgboost(ymat, xmat, ...)
}
\arguments{
\item{ymat}{data.frame or matrix. Dependent variable.}

\item{xmat}{data.frame or matrix. Independent variables.}

\item{...}{Arguments passed to \code{\link[xgboost:xgb.train]{xgboost::xgboost}}}
}
\value{
xgboost object
}
\description{
Fit XGBoost model
}
\author{
Insang Song
}
