% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_learner.R
\name{base_learner_prep}
\alias{base_learner_prep}
\title{Data preparation for base learners}
\usage{
base_learner_prep(
  learner = c("cnn", "randomforest", "xgboost"),
  data,
  dependent_name = "pm2.5",
  independent_name
)
}
\arguments{
\item{learner}{character(1). One of 'cnn', 'randomforest', 'xgboost'}

\item{data}{stdt. see \code{\link{convert_stobj_to_stdt}}}

\item{dependent_name}{Name of dependent variable. Default is "pm2.5"}

\item{independent_name}{character. Names of independent variables.}
}
\value{
A list of two matrices (except for cnn) or
multidimensional arrays (cnn) depending on learners
}
\description{
Data preparation for base learners
}
\author{
Insang Song
}
