% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_pipeline_example.R
\name{build_pipeline}
\alias{build_pipeline}
\title{Title Build_Pipeline}
\usage{
build_pipeline(
  base_learner_list,
  base_attr_list,
  meta_learner,
  response,
  covariate,
  obs_locs,
  pred_locs
)
}
\arguments{
\item{base_learner_list}{}

\item{base_attr_list}{}

\item{meta_learner}{}

\item{response}{}

\item{covariate}{}

\item{obs_locs}{}

\item{pred_locs}{}
}
\value{
put it all together
}
\description{
Title Build_Pipeline
}
\examples{
NULL
}
