% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_covariates.R
\name{calc_covariates}
\alias{calc_covariates}
\title{Calculate covariates}
\usage{
calc_covariates(
  covariate = c("modis", "MODIS", "koppen-geiger", "koeppen-geiger", "koppen", "koeppen",
    "geos", "dummies", "gmted", "roads", "sedac_groads", "mlcd", "tri", "ncep", "aadt",
    "ecoregions", "ecoregion"),
  path = "./input/koppen_geiger/raw/Beck_KG_V1_present_0p0083.tif",
  sites,
  id_col = "site_id",
  ...
)
}
\arguments{
\item{covariate}{character(1). Covariate type.}

\item{path}{character. Single or multiple path strings.}

\item{sites}{sf/SpatVector. Unique sites. Should include
a unique identifier field named \code{id_col}}

\item{id_col}{character(1). Name of unique identifier.
Default is \code{'site_id'}.}

\item{...}{Arguments passed to each covariate calculation
function.}
}
\value{
Calculated covariates. Mainly data.frame object.
}
\description{
Calculate covariates
}
\seealso{
- \link{calc_modis}: "modis", "MODIS"
- \link{calc_koppen_geiger}: "koppen-geiger", "koeppen-geiger", "koppen",
"koeppen"
}
\author{
Insang Song
}
