% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_base_functions.R
\name{calc_gmted_direct}
\alias{calc_gmted_direct}
\title{Reflown gmted processing}
\usage{
calc_gmted_direct(
  variable = NULL,
  path = NULL,
  locs = NULL,
  locs_id = NULL,
  win = c(-126, -62, 22, 52),
  radius = 0,
  fun = "mean",
  ...
)
}
\description{
Reflown gmted processing
}
