% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_base_functions.R
\name{calc_narr2}
\alias{calc_narr2}
\title{Calculate aggregated values for specified locations}
\usage{
calc_narr2(from, locs, locs_id = NULL, radius = 0, fun = "mean", ...)
}
\arguments{
\item{from}{The raster dataset from which to extract values.}

\item{locs}{A data frame containing the locations for which
to calculate aggregated values.
It should have a column in \code{locs_id} value
that contains unique identifiers for each location.}

\item{locs_id}{An optional column name
in the \code{locs} data frame that contains additional location
identifiers.}

\item{radius}{The radius within which to include neighboring locations
for aggregation. Default is 0.}

\item{fun}{The aggregation function to use.
It can be a character string specifying a function name
(e.g., "mean", "sum"),
or it can be a custom function. Default is "mean".}

\item{...}{Additional arguments to be passed to
the aggregation function.}
}
\value{
A data frame containing the aggregated values for each
location and time point.
}
\description{
This function calculates aggregated values for specified locations from
a raster dataset.
}
\seealso{
Other Calculation: 
\code{\link{calc_geos_strict}()},
\code{\link{calc_gmted_direct}()},
\code{\link{calculate}()},
\code{\link{inject_calculate}()},
\code{\link{inject_geos}()},
\code{\link{inject_gmted}()},
\code{\link{inject_modis_par}()},
\code{\link{par_narr}()},
\code{\link{process_counties}()},
\code{\link{process_geos_bulk}()},
\code{\link{process_narr2}()}
}
\concept{Calculation}
