% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_nlcd_ratio.R
\name{calc_nlcd_ratio}
\alias{calc_nlcd_ratio}
\title{Compute land cover classes ratio in circle buffers around points}
\usage{
calc_nlcd_ratio(data_vect, buf_radius = 1000, year = 2021, nlcd_path)
}
\arguments{
\item{data_vect}{terra::SpatVector of points geometry}

\item{buf_radius}{numeric (non-negative) giving the
radius of buffer around points}

\item{year}{numeric giving the year of NLCD data used}

\item{nlcd_path}{character giving nlcd data path}
}
\description{
Compute land cover classes ratio in circle buffers around points
}
