% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calc_tri_mod}
\alias{calc_tri_mod}
\title{Modified isotropic Sum of Exponentially Decaying Contributions (SEDC)
covariates}
\usage{
calc_tri_mod(
  from = NULL,
  locs,
  locs_id = "site_id",
  radius = c(1000L, 10000L, 50000L),
  geom = FALSE,
  ...
)
}
\arguments{
\item{from}{SpatVector(1). Output of \code{process_tri()}.}

\item{locs}{sf/SpatVector. Locations where TRI variables are calculated.}

\item{locs_id}{character(1). Unique site identifier column name.
Default is \code{"site_id"}.}

\item{radius}{Circular buffer radius.
Default is \code{c(1000, 10000, 50000)} (meters)}

\item{geom}{FALSE/"sf"/"terra".. Should the function return with geometry?
Default is \code{FALSE}, options with geometry are "sf" or "terra". The
coordinate reference system of the \code{sf} or \code{SpatVector} is that of \code{from.}}

\item{...}{Placeholders.}
}
\value{
a data.frame or SpatVector object
}
\description{
Calculate toxic release values for polygons or isotropic buffer point
locations. Returns a \code{data.frame} object containing \code{locs_id}
and variables for each chemical in \code{from}.
}
\note{
U.S. context.
}
\seealso{
\code{\link{sum_edc}}, \code{\link{process_tri}}
}
\author{
Insang Song, Mariana Kassien
}
