% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate_modis_direct}
\alias{calculate_modis_direct}
\title{Calculate MODIS from preprocessed file}
\usage{
calculate_modis_direct(file, site, site_id, radius, colheader, mark = TRUE)
}
\arguments{
\item{file}{character(1). File path of preprocessed GeoTIFF file.}

\item{site}{data.frame/sf. AQS sites. data.frame will be converted to sf.}

\item{site_id}{character(1). Unique identifier of \code{site}}

\item{radius}{numeric(1). Circular buffer radius. Distance unit is inherited
from the sites' coordinate system.}

\item{colheader}{character(1). Column name header.}

\item{mark}{logical(1). Mark zero-padded (length of 5) radius in the column name.}
}
\description{
Calculate MODIS from preprocessed file
}
