% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{cancel}
\alias{cancel}
\title{Cancel current workflow.}
\usage{
cancel(job = NULL)
}
\arguments{
\item{job}{character(1). SLURM job number or "full", to cancel all jobs
running on \verb{$USER} SLURM profile.}
}
\value{
NULL; cancels jobs.
}
\description{
Cancel current workflow.
}
\keyword{Miscellaneous}
