% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_learner.R
\name{check_and_load_torch}
\alias{check_and_load_torch}
\title{Check torch installation and load}
\usage{
check_and_load_torch(default_device = c("cpu", "cuda", "mps"))
}
\arguments{
\item{default_device}{character(1). "cpu" or "cuda"}
}
\description{
Check torch installation and load
}
\author{
Insang Song
}
