% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_outputs.R
\name{check_crs_is_valid}
\alias{check_crs_is_valid}
\title{Check if the output is with the valid coordinate reference system}
\usage{
check_crs_is_valid(model_output, crs_list = c("EPSG:4326", "EPSG:5070"))
}
\arguments{
\item{model_output}{sf/sftime object of model output.}

\item{crs_list}{a character/integer vector of acceptable CRS.
Default is c("EPSG:4326", "EPSG:5070")}
}
\value{
A logical value indicating the model is compliant to one of
 elements in crs_list.
}
\description{
Check if the output is with the valid coordinate reference system
}
\author{
Insang Song
}
