% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_outputs.R
\name{check_data_completeness}
\alias{check_data_completeness}
\title{Check if the output covariates are complete}
\usage{
check_data_completeness(
  model_output,
  fields_to_check,
  report_fields_na = FALSE
)
}
\arguments{
\item{model_output}{sf/sftime object of model output.}

\item{fields_to_check}{character(varying).
Field names where completeness will be checked.}

\item{report_fields_na}{logical(1). If any fields have NA values,
report the field names which gave the errors. Default is FALSE.}
}
\value{
A logical (when report_fields_na is FALSE;
TRUE means there are NA values at least one data value) or a list
(when report_fields_na is TRUE) object.
A list includes the list of fields that include NAs.
}
\description{
Check if the output covariates are complete
}
\author{
Insang Song
}
