% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_inputs.R
\name{check_input_is_stdt}
\alias{check_input_is_stdt}
\title{Type check of internal class stdt}
\usage{
check_input_is_stdt(input)
}
\arguments{
\item{input}{stdt object, which is a result of convert_stobj_to_stdt}
}
\value{
logical(1). TRUE (stdt is in one of class names) or
FALSE (otherwise)
}
\description{
Detects the class whether it is subject to stdt.
}
\author{
Insang Song
}
