% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_inputs.R
\name{check_input_raster_in_extent}
\alias{check_input_raster_in_extent}
\title{Check if the input raster is in the expected extent}
\usage{
check_input_raster_in_extent(
  input_raster,
  spatial_domain,
  domain_tolerance = 300000L
)
}
\arguments{
\item{input_raster}{SpatRaster/stars object}

\item{spatial_domain}{sf/sftime/SpatVector object of spatial domain}

\item{domain_tolerance}{numeric(1). Extrusion (in meters) from the bbox
of spatial_domain.}
}
\value{
A numeric value reporting the extent of input_raster is
within the extent of spatial_domain
}
\description{
\code{domain_tolerance} will have "meters" is converted to arc degrees approximately 100,000 meters = 1 arc degree
}
\author{
Kyle Messier, Insang Song
}
