% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_outputs.R
\name{check_means_are_valid}
\alias{check_means_are_valid}
\title{Check if the output prediction mean values are in the valid range}
\usage{
check_means_are_valid(
  model_output,
  model_mean_name = "prediction_mean",
  observation,
  observation_mean_name = "value",
  tolerance_factor = 3
)
}
\arguments{
\item{model_output}{sf/sftime object of model output}

\item{model_mean_name}{character(1). the name of layer where mean values are stored}

\item{observation}{a data.frame with observations}

\item{observation_mean_name}{character(1). field name of observations in observation object.}

\item{tolerance_factor}{numeric(1). denominator(min) multiplier(max)}
}
\value{
Logical value indicating the mean values are within the range or not.
}
\description{
Check if the output prediction mean values are in the valid range
}
\author{
Insang Song
}
