% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_outputs.R
\name{check_output_locs_are_valid}
\alias{check_output_locs_are_valid}
\title{Check output locations are in the reference spatial domain of the mainland US}
\usage{
check_output_locs_are_valid(model_output, spatial_domain)
}
\arguments{
\item{model_output}{sf/sftime object of the model output}

\item{spatial_domain}{sf/sftime object of spatial domain}
}
\value{
A logical vector of length nrow(model_output)
}
\description{
Check output locations are in the reference spatial domain of the mainland US
}
\author{
Insang Song
}
