% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_temporal_range.R
\name{check_temporal_range}
\alias{check_temporal_range}
\title{Check if input data is within user defined temporal range}
\usage{
check_temporal_range(
  data = NULL,
  start_range = "2018-01-01",
  end_range = "2022-12-31",
  data_type = NULL,
  timezone = "UTC",
  time_column = "Date"
)
}
\arguments{
\item{data}{file path to data}

\item{start_range}{beginning of temporal range}

\item{end_range}{end of temporal range}

\item{data_type}{Point, Polygon, or Raster data}

\item{timezone}{character.
Time zone in three uppercase letters (e.g., UTC, EST, etc.)}

\item{time_column}{character. Name of the column storing
time information.}
}
\value{
A logical vector reporting whether time observations in data are
within defined temporal range
}
\description{
Check if input data is within user defined temporal range
}
\author{
Mitchell Manware
}
