% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_test_support.R
\name{check_url_file_exist}
\alias{check_url_file_exist}
\title{Check if sample of download URLs have HTTP Status 200}
\usage{
check_url_file_exist(url, method = "HEAD")
}
\arguments{
\item{url}{Download URL to be checked.}

\item{method}{httr method to obtain URL ("HEAD" or "GET")}
}
\description{
Check if sample of download URLs have HTTP Status 200
}
\author{
Insang Song; Mitchell Manware
}
