% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_outputs.R
\name{check_variances_are_valid}
\alias{check_variances_are_valid}
\title{Check if the output prediction variances are non-negative}
\usage{
check_variances_are_valid(
  model_output,
  model_variance_name = "prediction_variance"
)
}
\arguments{
\item{model_output}{sf/sftime object of model output}

\item{model_variance_name}{character(1). the name of layer where variance
values are stored}
}
\value{
Logical value indicating the variance are valid (TRUE) or not.
}
\description{
Check if the output prediction variances are non-negative
}
\author{
Kyle P Messier
}
