% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_base_functions.R
\name{df_params}
\alias{df_params}
\title{Get data.frame of function parameters}
\usage{
df_params(functions)
}
\arguments{
\item{functions}{character. Vector of function names.}
}
\value{
A data.frame containing the parameters of the functions.
}
\description{
Get data.frame of function parameters
}
\seealso{
Other Pipeline utility: 
\code{\link{attach_xy}()},
\code{\link{feature_raw_download}()},
\code{\link{load_modis_files}()},
\code{\link{read_locs}()},
\code{\link{read_paths}()},
\code{\link{search_function}()},
\code{\link{set_slurm_resource}()}
}
\concept{Pipeline utility}
