% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_xgboost_data.R
\name{double_to_site_id}
\alias{double_to_site_id}
\title{Create xgboost data
Function used in merge_covariates to convert site_id
from double to character}
\usage{
double_to_site_id(x)
}
\arguments{
\item{x}{A double (of length 1 with 14 digits or less)}
}
\value{
A character with 14 digits (site_id format)
}
\description{
Create xgboost data
Function used in merge_covariates to convert site_id
from double to character
}
\author{
Eva Marques
}
