% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_data}
\alias{download_data}
\title{download_data: a wrapper function for download covariate sources}
\usage{
download_data(
  dataset_name = c("aqs", "ecoregion", "geos", "gmted", "koppen", "koppengeiger",
    "merra2", "merra", "narr_monolevel", "modis", "narr_p_levels", "nlcd", "noaa",
    "sedac_groads", "sedac_population", "groads", "population", "plevels", "p_levels",
    "monolevel", "hms", "smoke"),
  directory_to_save = NULL,
  data_download_acknowledgement = FALSE,
  ...
)
}
\arguments{
\item{dataset_name}{character(1). Dataset to download.}

\item{directory_to_save}{character(1). Directory to save / unzip
(if zip files are downloaded) data.}

\item{data_download_acknowledgement}{logical(1). By setting `= TRUE` the
user acknowledge that the data downloaded using this function may be very
large and use lots of machine storage and memory.}

\item{...}{Arguments passed to each download function.}
}
\description{
download_data: a wrapper function for download covariate sources
}
\note{
- All download function names are in \code{download_*_data} formats
}
\seealso{
- [download_aqs_data()]: "aqs"
- [download_ecoregion_data()]: "ecoregion"
- [download_geos_cf_data()]: "geos"
- [download_gmted_data()]: "gmted"
- [download_koppen_geiger_data()]: "koppen", "koppengeiger"
- [download_merra2_data()]: "merra2", "merra"
- [download_narr_monolevel_data()]: "narr_monolevel", "monolevel"
- [download_narr_p_levels_data()]: "narr_p_levels", "p_levels", "plevels"
- [download_nlcd_data()]: "nlcd",
- [download_noaa_hms_smoke_data()]: "noaa", "smoke", "hms"
- [download_sedac_groads_data()]: "sedac_groads", "groads"
- [download_sedac_population_data()]: "sedac_population", "population"
- [download_modis_data()]: "modis", "MODIS"
}
\author{
Insang Song
}
