% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_ecoregion_data}
\alias{download_ecoregion_data}
\title{download_ecoregion_data: download Ecoregion Shapefiles from EPA.}
\usage{
download_ecoregion_data(
  directory_to_download = "./input/ecoregions/",
  directory_to_save = "./input/ecoregions/",
  data_download_acknowledgement = FALSE,
  unzip = TRUE,
  remove_zip = FALSE,
  download = FALSE,
  remove_command = TRUE
)
}
\arguments{
\item{directory_to_download}{character(1). Directory to download zip file
of Ecoregion level 3 shapefiles}

\item{directory_to_save}{character(1). Directory to decompress zip files.}

\item{data_download_acknowledgement}{logical(1). By setting \code{TRUE} the
user acknowledge that the data downloaded using this function may be very
large and use lots of machine storage and memory.}

\item{unzip}{logical(1). Unzip zip files. Default = `TRUE`.}

\item{remove_zip}{logical(1). Remove zip file from directory_to_download.
Default = `FALSE`.}

\item{download}{logical(1). `= FALSE` will generate a `.txt` file containing
all download commands. By setting `= TRUE` the function will download all of
the requested data files. Default is FALSE.}

\item{remove_command}{logical(1). Remove (\code{TRUE}) or keep (\code{FALSE})
the text file containing download commands.}
}
\value{
NULL;
}
\description{
The \code{download_ecoregion_data()}
function accesses and downloads Ecoregions
level 3 data, where all pieces of information in the higher levels are
included.
}
\author{
Insang Song
}
