% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_support.R
\name{download_run}
\alias{download_run}
\title{download_run: execute or skip `system_command` in data download function.}
\usage{
download_run(download = FALSE, system_command = NULL, commands_txt = NULL)
}
\arguments{
\item{download}{logical(1). Execute (`TRUE`) or skip (`FALSE`) download.}

\item{system_command}{character(1). Linux command to execute downloads.
Inherited from data download function.}

\item{commands_txt}{character(1).}
}
\description{
Execute or skip the commands listed in the ...wget/curl_commands.txt file
produced by one of the data download functions.
}
