% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_sedac_groads_data}
\alias{download_sedac_groads_data}
\title{download_sedac_groads_data: download Global Roads Open Access Data
Set (gROADS), v1 (1980-2010) data from NASA Socioeconomic Data and
Applications Center (SEDAC).}
\usage{
download_sedac_groads_data(
  data_format = c("Shapefile", "Geodatabase"),
  data_region = c("Americas", "Global", "Africa", "Asia", "Europe", "Oceania East",
    "Oceania West"),
  directory_to_download = "./input/sedac_groads/",
  directory_to_save = "./input/sedac_groads/",
  data_download_acknowledgement = FALSE,
  unzip = TRUE,
  remove_zip = FALSE,
  download = FALSE,
  remove_command = FALSE
)
}
\arguments{
\item{data_format}{character(1). Data can be downloaded as "Shapefile" or
"Geodatabase". (Only "Geodatabase" available for "Global" region).}

\item{data_region}{character(1). Data can be downloaded for "Global",
"Africa", "Asia", "Europe", "Americas", "Oceania East", and "Oceania West".}

\item{directory_to_download}{character(1). Directory to download zip files
from NASA UN WPP-Adjusted Population Density, v4.11.}

\item{directory_to_save}{character(1). Directory to decompress zip files.}

\item{data_download_acknowledgement}{logical(1). By setting `= TRUE` the user
acknowledge that the data downloaded using this function may be very large
and use lots of machine storage and memory.}

\item{unzip}{logical(1). Unzip zip files. Default = `TRUE`.}

\item{remove_zip}{logical(1). Remove zip files from directory_to_download.
Default = `FALSE`.}

\item{download}{logical(1). `= FALSE` will generate a `.txt` file containing
all download commands. By setting `= TRUE` the function will download all of
the requested data files.}

\item{remove_command}{logical(1). Remove (\code{TRUE}) or keep (\code{FALSE})
the text file containing download commands.}
}
\value{
NULL;
}
\description{
The `download_sedac_groads_data()` function accesses and downloads
roads data from the National Aeronautics and Space
Administration's (NASA) [Global Roads Open Access Data Set]
(https://sedac.ciesin.columbia.edu/data/set/groads-global-roads-open-access-
v1/data-download).
}
\author{
Mitchell Manware, Insang Song
}
