% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_spacetime_data.R
\name{dt_to_sftime}
\alias{dt_to_sftime}
\title{Create a sftime object from a data.table}
\usage{
dt_to_sftime(datatable, crs)
}
\arguments{
\item{datatable}{A data.table object with columns "lat", "lon", "date"}

\item{crs}{A character containing the original crs}
}
\value{
an sftime object
}
\description{
Create a sftime object from a data.table
}
\author{
Eva Marques
}
