% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{export_tif}
\alias{export_tif}
\title{Export preprocessed GeoTIFF file of one day MODIS product files}
\usage{
export_tif(path_in, product_code, pat, subdataset = "", dest = ".")
}
\arguments{
\item{path_in}{character. File paths of MODIS product files on a day}

\item{product_code}{character(1). Product code. e.g., "VNP46A2".}

\item{pat}{character. Filter of \code{path_in}.}

\item{subdataset}{character. Subdatasets to merge. It is ignored
when \code{product_code} is one of \code{"MOD06_L2"} or \code{"VNP46A2"}.}

\item{dest}{character(1). Directory path to export the result file.}
}
\value{
A GeoTIFF file is saved to
\verb{\{dest\}/\{product_code\}_processed_\{date\}.tif} and this path is returned.
}
\description{
Export preprocessed GeoTIFF file of one day MODIS product files
}
\examples{
\dontrun{
export_tif(c("A2018001.001.hdf", "A2018001.002.hdf"),
           product_code = "MOD11A1",
           pat = "A\\\\d{7,7}",
           subdataset = "16 day NDVI",
           dest = ".")
# mirai run
library(mirai)
basedir61 <- "/mnt/modis/raw/61"
product <- "MCD19A2"
paths <-
  list.files(
    file.path(basedir61, product),
    pattern = "*.*.hdf$",
    full.names = TRUE,
    recursive = TRUE
  )
pcode <- "MCD19A2"
pattern <- "A\\\\d{7,7}"
subdataset <- c("Optical_Depth_055")
dest = "mydir"

run_mirai <-
  mirai_map(
    paths,
    .f = export_tif,
    .args = list(
       product_code = product,
       pat = pattern,
       subdataset = subdataset,
       dest = dest
    )
  )
run_mirai[.progress]
}
}
\author{
Insang Song
}
