% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_base_functions.R
\name{fit_base}
\alias{fit_base}
\title{Wrapper for base learners}
\usage{
fit_base(learner, rset, full_data, ...)
}
\arguments{
\item{learner}{character(1). One of "mlp", "xgboost", "elasticnet"}

\item{rset}{pre-generated rset object.}

\item{full_data}{data.frame. The imputed full data that will be used to
restore full rset object for model fitting.}

\item{...}{Additional arguments to be passed.}
}
\value{
The fitted workflow.
}
\description{
Wrapper for base learners
}
\keyword{Baselearner}
