% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_learner.R
\name{fit_base_cv_switcher}
\alias{fit_base_cv_switcher}
\title{Choose cross-validation strategy for the base learner}
\usage{
fit_base_cv_switcher(learner = c("spatial", "temporal", "spatiotemporal"), ...)
}
\arguments{
\item{learner}{character(1). Learner type.
\itemize{
\item "spatial": spatial cross-validation.
\item "temporal": temporal cross-validation.
\item "spatiotemporal": spatiotemporal cross-validation.
}}

\item{...}{Additional arguments to be passed.}
}
\value{
rsample::manual_rset() object.
}
\description{
Choose cross-validation strategy for the base learner
}
\note{
This function's returned value is used as an input for
\code{fit_base_brulee}, \code{fit_base_lightgbm}, and \code{fit_base_elnet}.
Learner values can be used as a branching point for the cross-validation
strategy.
}
\keyword{Baselearner}
