% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_learner.R
\name{fit_base_learner_cnn}
\alias{fit_base_learner_cnn}
\title{Fit convolutional neural networks with neural network library}
\usage{
fit_base_learner_cnn(
  covars,
  dependent_name,
  independent_name,
  cv_mode,
  cv_fold
)
}
\arguments{
\item{covars}{stdt. See \link{convert_stobj_to_stdt}}

\item{dependent_name}{character(1). Name of the dependent variable.}

\item{independent_name}{character(1). Names of independent variables.}

\item{cv_mode}{character(1). One of 'lolo (leave-one-location-out)',
'loto (leave-one-time-out)',
'lolto (leave-one-location-time-out)',
'lblo (leave-block-location-out)',
'lbto (leave-block-time-out)',
'random (full random selection)'}

\item{cv_fold}{integer(1). Number of folds for cross-validation.}
}
\value{
data.frame
}
\description{
This function uses torch as a backend.
Torch currently supports CPU, CUDA, and Metal (Apple Silicon graphics),
thus users can leverage available computational assets
to accelerate the model fitting.
}
\author{
Insang Song
}
