% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_learner.R
\name{fit_base_tune}
\alias{fit_base_tune}
\title{Tune base learner}
\usage{
fit_base_tune(
  recipe,
  model,
  resample,
  tune_mode = c("bayes", "grid"),
  grid = NULL,
  iter_bayes = 10L,
  trim_resamples = TRUE,
  return_best = TRUE
)
}
\arguments{
\item{recipe}{The recipe object.}

\item{model}{The model object.}

\item{resample}{The resample object.}

\item{tune_mode}{character(1). Hyperparameter tuning mode.
Default is "bayes", "grid" is acceptable.}

\item{grid}{The grid object for hyperparameter tuning.}

\item{trim_resamples}{logical(1). Default is TRUE, which replaces the actual
data.frames in splits column of \code{tune_results} object with NA.}

\item{return_best}{logical(1). If TRUE, the best tuned model is returned.}
}
\value{
List of 2:
\itemize{
\item \code{base_prediction}: \code{data.frame} of the best model prediction.
\item \code{base_parameter}: \code{tune_results} object of the best model.
}
}
\description{
Tune base learner
}
\keyword{Baselearner}
\keyword{internal}
