% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_learner.R
\name{fit_meta_learner}
\alias{fit_meta_learner}
\title{Fit meta learner}
\usage{
fit_meta_learner(
  data,
  p_col_sel = 0.5,
  rset = NULL,
  yvar = "pm2.5",
  xvar = character(0),
  tune_iter = 50L
)
}
\arguments{
\item{data}{data.frame. Full data.}

\item{p_col_sel}{numeric(1). Rate of column resampling. Default is 0.5.}

\item{rset}{rset object. Specification of training/test sets.}

\item{yvar}{character(1). Outcome variable name}

\item{xvar}{character. Feature names.}

\item{tune_iter}{integer(1). Bayesian optimization iterations.}
}
\value{
best model object.
}
\description{
This function subsets the full data by column subsamples (rate=50\%)
The optimal hyperparameter search is performed based on spatial,
temporal, and spatiotemporal cross-validation schemes and continuous
ranked probability score (CRPS) with Bayesian optimization.
}
