% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_learner.R
\name{fit_prediction}
\alias{fit_prediction}
\title{Predict base learners at for cross validation folds.}
\usage{
fit_prediction(
  fit = NULL,
  test = NULL,
  target_cols = c("site_id", "time", "lon", "lat"),
  name = NULL
)
}
\arguments{
\item{fit}{list(1). List with 1) trained workflow and 2) performance metrics.
Workflow must be first item in list.}

\item{test}{data.table(1). \code{data.table} with un-trained testing data.}

\item{target_cols}{characters(1). Columns to retain from the full test data
`data.frame``.}

\item{name}{character(1). Name for prediction column. Should identify the
model and engine and/or hyperparameters.}
}
\value{
\code{data.table} with attached site identifiers
}
\description{
Predict base learners at for cross validation folds.
}
\keyword{meta_learner}
