% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{fl_dates_flatten}
\alias{fl_dates_flatten}
\title{Extract the first and last elements of a list of date vectors}
\usage{
fl_dates_flatten(dates)
}
\arguments{
\item{dates}{list. A list of dates.}
}
\value{
a character vector with the first and last dates from the list.
}
\description{
It flattens the list first, then extracts the first and the last dates.
}
\keyword{Utility}
