% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{generate_block_sp_index}
\alias{generate_block_sp_index}
\title{Generate blocked spatial index}
\usage{
generate_block_sp_index(covars, cv_fold = NULL, blocks = NULL, block_id = NULL)
}
\arguments{
\item{covars}{stdt.}

\item{cv_fold}{integer(1). Number of folds for cross-validation.}

\item{blocks}{numeric(2)/sf/SpatVector configuration of blocks.}

\item{block_id}{character(1). The unique identifier of each block.}
}
\value{
stdt
}
\description{
Generate blocked spatial index
}
\details{
"Block" in this function refers to a group of contiguous spatial/temporal entities.
Regardless of mode values (should be one of 'spatial' or 'spatiotemporal'),
the index column will be named 'sp_index'. \code{block_id} should be set with a proper field name
when blocks is a sf or SpatVector object. When \code{cv_fold} is an integer, then the coordinates stored in \code{stdt} object are clustered with k-means. In this case, the result object will include attributes (accessible with \code{attr} function) about
"kmeans_centers" (coordinates of centers) and "kmeans_sizes" (size of each cluster)
}
\author{
Insang Song
}
