% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{generate_cv_index_lblto}
\alias{generate_cv_index_lblto}
\title{Generate spatio-temporal cross-validation index
(leave-block-location-time-out)}
\usage{
generate_cv_index_lblto(covars, sp_fold, t_fold, blocks, block_id = NULL)
}
\arguments{
\item{covars}{stdt. See \link[NRTAPModel]{convert_stobj_to_stdt}
for details.}

\item{sp_fold}{integer(1). Number of subfolds for spatial blocks.}

\item{t_fold}{integer(1). Number of subfolds for temporal blocks.}

\item{blocks}{integer(2)/sf/SpatVector object.}

\item{block_id}{character(1). The unique identifier of each block.}
}
\value{
An integer vector.
}
\description{
Generate spatio-temporal cross-validation index
(leave-block-location-time-out)
}
\details{
The maximum of results is \code{sp_fold * t_fold}.
}
\author{
Insang Song
}
