% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{generate_cv_index_lbto}
\alias{generate_cv_index_lbto}
\title{Generate spatio-temporal cross-validation index (leave-block-time-out)}
\usage{
generate_cv_index_lbto(covars, cv_fold = NULL)
}
\arguments{
\item{covars}{stdt. See \code{\link{convert_stobj_to_stdt}} for details.}

\item{cv_fold}{integer(1). Number of folds for cross-validation.}
}
\value{
An integer vector.
}
\description{
Generate spatio-temporal cross-validation index (leave-block-time-out)
}
\author{
Insang Song
}
