% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{generate_cv_index_random}
\alias{generate_cv_index_random}
\title{Generate spatio-temporal cross-validation index
(random)}
\usage{
generate_cv_index_random(covars, cv_fold = NULL)
}
\arguments{
\item{covars}{stdt. See \link[NRTAPModel]{convert_stobj_to_stdt}
for details.}

\item{cv_fold}{integer(1). Number of folds for cross-validation.}
}
\value{
An integer vector with unique values of \code{seq(1, cv_fold)}
}
\description{
Generate spatio-temporal cross-validation index
(random)
}
\author{
Insang Song
}
