% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_learner.R
\name{generate_cv_ts}
\alias{generate_cv_ts}
\title{Generate temporal cross-validation index}
\usage{
generate_cv_ts(data, time_col = "time", cv_fold = 10L, window = 90)
}
\arguments{
\item{data}{data.table with X, Y, and time information.}

\item{time_col}{character(1). Field name with time information.}

\item{cv_fold}{integer(1). Number of cross-validation folds.}

\item{window}{integer(1). Window size for each fold. Unit is
the base unit of temporal values stored in \code{time_col}.}
}
\value{
rsample::manual_rset() object.
}
\description{
Generate temporal cross-validation index
}
\details{
\code{window} is used to determine the padding or skipping
intervals for temporal cross-validation. Skipping intervals are
needed not to training and testing data to overlap.
}
\keyword{Baselearner}
\keyword{development}
