% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{generate_spt_index}
\alias{generate_spt_index}
\title{Generate unique spatiotemporal identifier from stdt object}
\usage{
generate_spt_index(covars, mode = c("spatial", "spatiotemporal"))
}
\arguments{
\item{covars}{stdt.}

\item{mode}{One of "spatial" or "spatiotemporal"}
}
\value{
stdt with a new index named "sp_index"
}
\description{
It generates unique spatiotemporal identifier in the
input stdt object. Regardless of mode values
(should be one of 'spatial' or 'spatiotemporal'),
the index column will be named 'sp_index'.
}
\author{
Insang Song
}
