% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_pipeline_example.R
\name{generic_base_learner}
\alias{generic_base_learner}
\title{define a base learner dummy: this illustrates one way to wrap a learning
algorithm and return an object that makes predictions}
\usage{
generic_base_learner(response, covariate, obs_locs, model_attr)
}
\arguments{
\item{response}{}

\item{covariate}{}

\item{obs_locs}{}

\item{model_attr}{}
}
\description{
define a base learner dummy: this illustrates one way to wrap a learning
algorithm and return an object that makes predictions
}
