% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_pipeline_example.R
\name{generic_meta_learner}
\alias{generic_meta_learner}
\title{Title Generic_Meta_Learner}
\usage{
generic_meta_learner(base_predictor_list, pred_loc)
}
\arguments{
\item{base_predictor_list}{list of the base predictor output}

\item{pred_loc}{geospatial information on the locations for predicting}
}
\value{
mean and variance of the meta learner at prediction locations
}
\description{
Title Generic_Meta_Learner
}
\examples{
NULL
}
