% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_base_functions.R
\name{get_aqs_data}
\alias{get_aqs_data}
\title{Filter monitors with the minimum POC value}
\usage{
get_aqs_data(
  path = list.files(path = meta_run("dir_input_aqs"), pattern =
    "daily_88101_[0-9]{4}.csv", full.names = TRUE),
  site_spt = NULL,
  locs_id = meta_run("char_siteid"),
  time_id = meta_run("char_timeid"),
  poc_name = "POC",
  sampling = "Sample.Duration",
  date_start = "2018-01-01",
  date_end = "2022-12-31"
)
}
\arguments{
\item{path}{data.frame/tibble/data.table}

\item{site_spt}{Space-time site data.}

\item{locs_id}{character(1). Name of site id (not monitor id)}

\item{poc_name}{character(1). Name of column containing POC values.}

\item{sampling}{character(1). Name of column with sampling duration.}

\item{date_start}{character(1).}

\item{date_end}{character(1).}
}
\value{
a data.table object
}
\description{
Filter monitors with the minimum POC value
}
\author{
Insang Song
}
