% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_base_functions.R
\name{\%tin\%}
\alias{\%tin\%}
\title{Check if a query date falls within a time interval}
\usage{
query_date \%tin\% tvec
}
\arguments{
\item{query_date}{The query date to check.}

\item{tvec}{A vector of two dates defining the time interval.}
}
\value{
TRUE if the query date falls within the time interval,
FALSE otherwise.
}
\description{
This function checks if a given query date falls within a time interval
defined by a vector of two dates.
}
\examples{
query_date <- as.Date("2022-01-01")
tvec <- c(as.Date("2021-01-01"), as.Date("2023-01-01"))
`\%tin\%`(query_date, tvec)

}
\concept{Miscellaneous}
