% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_base_functions.R
\name{impute_all}
\alias{impute_all}
\title{Impute missing values and attach lagged features}
\usage{
impute_all(
  dt,
  period,
  nthreads_dt = 32L,
  nthreads_collapse = 32L,
  nthreads_imputation = 32L
)
}
\arguments{
\item{dt}{The input data table to be imputed.}

\item{period}{The period for lagged features.}

\item{nthreads_dt}{The number of threads to be used for
data.table operations.}

\item{nthreads_collapse}{The number of threads to be used for
collapse operations.}

\item{nthreads_imputation}{The number of threads to be used for
the imputation process.}
}
\value{
The imputed data table with lagged features.
}
\description{
Impute missing values and attach lagged features
}
\note{
under construction.
This function performs imputation on a given data table
by replacing missing values with imputed values.
It follows a series of steps including data cleaning, name cleaning,
geoscf column renaming, NDVI 16-day backward filling,
zero-variance exclusion, excessive "true zeros" exclusion,
and imputation using missRanger.
}
\examples{
dt <- data.table(a = c(1, 2, NA, 4), b = c(NA, 2, 3, 4))
impute_all(dt, period = 1)

}
