% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_base_functions.R
\name{inject_geos}
\alias{inject_geos}
\title{Injects geographic information into a data frame}
\usage{
inject_geos(locs, injection)
}
\arguments{
\item{locs}{A data frame containing the locations for which
geographic information needs to be injected.}

\item{injection}{A list of additional arguments to be passed to
the \code{calc_geos_strict} function.}
}
\value{
A modified data frame with injected geographic information.
}
\description{
This function injects geographic information into a data frame using
the \code{calc_geos_strict} function. The injected information includes
latitude and longitude coordinates based on the specified locations,
a location ID column, a window range, and a snapping option.
}
\seealso{
Other Calculation: 
\code{\link{calc_geos_strict}()},
\code{\link{calc_gmted_direct}()},
\code{\link{calc_narr2}()},
\code{\link{inject_calculate}()},
\code{\link{inject_gmted}()},
\code{\link{inject_modis_par}()},
\code{\link{par_narr}()},
\code{\link{process_counties}()},
\code{\link{process_geos_bulk}()},
\code{\link{process_narr2}()}
}
\concept{Calculation}
