% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/injection.R
\name{inject_match}
\alias{inject_match}
\title{Injects the calculate function with matched arguments.}
\usage{
inject_match(f, args)
}
\arguments{
\item{f}{function.}

\item{args}{List of arguments that are attempted to be injected into \code{f}.}
}
\value{
Injected function evaluation.
}
\description{
Injects the calculate function with matched arguments.
}
\keyword{Calculation}
