% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/injection.R
\name{inject_modis_par}
\alias{inject_modis_par}
\title{Injects arguments to parallelize MODIS/VIIRS data processing}
\usage{
inject_modis_par(locs, injection)
}
\arguments{
\item{locs}{A data frame containing the locations for which MODIS
features need to be calculated.}

\item{injection}{\strong{List} of additional parameters to be passed to the
\code{calculate_modis} function.}
}
\value{
MODIS/VIIRS feature data.frame.
}
\description{
Injects arguments to parallelize MODIS/VIIRS data processing
}
\note{
Soon to be deprecated per dropping future dependency.
}
\examples{
\dontrun{
files <-
  c(
    "/downloads/modis/mod06/MOD06_L2.A2022001.0000.061.2022001160000.hdf",
   "/downloads/modis/mod06/MOD06_L2.A2022001.0005.061.2022001160000.hdf"
  )
my_locs <- data.frame(site_id = 1:2, lon = c(-88, -87), lat = c(35, 35))
my_locs <- sf::st_as_sf(my_locs, coords = c("lon", "lat"))
inject_modis_par(
  locs = my_locs,
  injection = list(path = files, subdataset = "Cloud_Fraction_Day",
     name_covariates = "MOD_CLCVD_0_",
     preprocess = amadeus::process_modis_swath, radius = c(1000)))
}
}
\seealso{
\code{\link[amadeus:calculate_modis_daily]{amadeus::calculate_modis_daily}}, \code{\link[amadeus:calculate_modis]{amadeus::calculate_modis}}
}
\keyword{Calculation}
