% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_point_receptor_to_polygon_source.r
\name{join_point_receptor_to_polygon_source}
\alias{join_point_receptor_to_polygon_source}
\title{join_point_receptor_to_polygon_source}
\usage{
join_point_receptor_to_polygon_source(
  receptor_sf = NULL,
  receptor_id = NULL,
  source_polygon_sf = NULL,
  source_polygon_id = NULL,
  add_all_to_output = TRUE,
  print_log_to_console = TRUE,
  write_log_to_file = TRUE
)
}
\arguments{
\item{receptor_sf}{}

\item{receptor_id}{}

\item{source_polygon_sf}{}

\item{source_polygon_id}{}

\item{add_all_to_output}{}

\item{print_log_to_console}{}

\item{write_log_to_file}{}
}
\value{
a data frame: receptor_point_to_source_polygon_crosswalk
}
\description{
join_point_receptor_to_polygon_source
}
\examples{
NULL
}
