% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{loadargs}
\alias{loadargs}
\title{Load arguments from the formatted argument list file}
\usage{
loadargs(argfile, dataset)
}
\arguments{
\item{argfile}{character(1). Path to the argument file. RDS format.}

\item{dataset}{character(1). Dataset name.}
}
\value{
A list of arguments stored in \code{dataset} slot of the
argument file.
}
\description{
This function loads the list object of arguments
to be injected into the calculation functions defined at
each target. The arguments are numeric or character, and some
of these are function names. In this case, the internal function
\code{unmarshal_function} is called to convert the function name
to the actual function.
}
\keyword{Utility}
