% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_learner.R
\name{make_subdata}
\alias{make_subdata}
\title{Make sampled subdataframes for base learners}
\usage{
make_subdata(data, n = NULL, p = 0.3, ngroup_init = NULL)
}
\arguments{
\item{data}{An object that inherits data.frame.}

\item{n}{The number of rows to be sampled.}

\item{p}{The proportion of rows to be used. Default is 0.3.}

\item{ngroup_init}{integer(1). Initial number of splits for
pairing groups. Default is NULL. Ensures that subsample is divisible
by \code{ngroup_init} for \code{generate_cv_index_spt}.}
}
\value{
The row index of the original data. The name of the original
data object is stored in attribute "object_origin".
}
\description{
Per beethoven resampling strategy, this function selects
the predefined number of rows from the input data table and
saves the row index in .rowindex field.
}
\keyword{Baselearner}
