% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_xgboost_data.R
\name{merge_covariates}
\alias{merge_covariates}
\title{Function to merge covariates in 3 separate of data.frames
\itemize{
\item one for only-spatial covariates (index = site_id)
\item one for yearly spatial covariates (index = site_id, year)
\item one for spatio-temporal covariates (index = site_id, time as days)
data.frames are saved as RDS files in the output folder
}}
\usage{
merge_covariates(covar_folder = "../output")
}
\arguments{
\item{covar_folder}{path character to folder where covariates are stored.
This function will read and merge all data stored in
NRTAP_Covars_*.rds files.}
}
\description{
Function to merge covariates in 3 separate of data.frames
\itemize{
\item one for only-spatial covariates (index = site_id)
\item one for yearly spatial covariates (index = site_id, year)
\item one for spatio-temporal covariates (index = site_id, time as days)
data.frames are saved as RDS files in the output folder
}
}
\author{
Eva Marques
}
