% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_covariates.R
\name{modis_get_vrt}
\alias{modis_get_vrt}
\title{Get mosaicked or merged raster from multiple MODIS hdf files}
\usage{
modis_get_vrt(
  paths,
  product = c("MOD11A1", "MOD13A2", "MOD06_L2", "MOD09GA", "MCD19A2"),
  index_sds = NULL,
  layers = NULL,
  date_in,
  foo = mean
)
}
\arguments{
\item{paths}{character. Full list of hdf file paths.
preferably a recursive search result from \code{list.files}.}

\item{product}{character(1). Product code of MODIS. Should be one of
\code{c('MOD11A1', 'MOD13A2', 'MOD06_L2', 'MOD09GA', 'MCD19A2')}}

\item{index_sds}{integer(1). The index of subdataset to draw.}

\item{layers}{integer. The index (indices) of layers if there are only
layers inside the input file.}

\item{date_in}{Date(1). date to query.}

\item{foo}{function. A function compatible with \code{SpatRaster}.}
}
\value{
A SpatRaster object.
}
\description{
Get mosaicked or merged raster from multiple MODIS hdf files
}
\author{
Insang Song
}
