% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_covariates.R
\name{modis_preprocess_vnp46}
\alias{modis_preprocess_vnp46}
\title{Assign MODIS VNP46 corner coordinates to retrieve a merged raster}
\usage{
modis_preprocess_vnp46(
  paths,
  date_in,
  tile_df,
  subdataset = 3L,
  crs_ref = "EPSG:4326"
)
}
\arguments{
\item{paths}{character. Full paths of h5 files.}

\item{date_in}{character(1). Date to query.}

\item{tile_df}{prespecified data.frame with "tile" and "exts" columns,
where the former stores tile number (h00v00 format) and the latter
stores terra::ext output.}

\item{subdataset}{integer(1). Subdataset number to process.
Default is 3L.}

\item{crs_ref}{character(1). terra::crs compatible CRS.
Default is "EPSG:4326"}
}
\description{
This function will return a SpatRaster object with
georeferenced h5 files of VNP46A2 product.
}
\author{
Insang Song
}
