% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_base_functions.R
\name{par_narr}
\alias{par_narr}
\title{Parallelize NARR feature calculation}
\usage{
par_narr(domain, date, locs, nthreads = 24L)
}
\arguments{
\item{domain}{A character vector specifying the domains to process.}

\item{date}{A character vector specifying the date of the NARR data to process.}

\item{locs}{A data frame specifying the locations to calculate NARR data for.}

\item{nthreads}{An integer specifying the number of threads to use for parallel processing. Default is 24.}
}
\value{
A list of results from the parallel processing.
}
\description{
This function parallelizes the processing and calculation of
NARR data for multiple domains.
}
\seealso{
Other Calculation: 
\code{\link{calc_geos_strict}()},
\code{\link{calc_gmted_direct}()},
\code{\link{calc_narr2}()},
\code{\link{calculate}()},
\code{\link{inject_calculate}()},
\code{\link{inject_geos}()},
\code{\link{inject_gmted}()},
\code{\link{inject_modis_par}()},
\code{\link{process_counties}()},
\code{\link{process_geos_bulk}()},
\code{\link{process_narr2}()}
}
\concept{Calculation}
