% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_base_functions.R
\name{post_calc_convert_time}
\alias{post_calc_convert_time}
\title{Convert time column to character}
\usage{
post_calc_convert_time(df)
}
\arguments{
\item{df}{data.table}
}
\description{
Convert time column to character
}
\note{
This function takes preprocessed data.table with
a column named \code{"time"}.
}
\seealso{
Other Post-calculation: 
\code{\link{add_time_col}()},
\code{\link{append_predecessors}()},
\code{\link{impute_all}()},
\code{\link{post_calc_autojoin}()},
\code{\link{post_calc_df_year_expand}()},
\code{\link{post_calc_join_yeardate}()},
\code{\link{post_calc_merge_all}()},
\code{\link{post_calc_merge_features}()},
\code{\link{post_calc_unify_timecols}()},
\code{\link{post_calc_year_expand}()},
\code{\link{reduce_merge}()}
}
\concept{Post-calculation}
