% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_base_functions.R
\name{post_calc_merge_features}
\alias{post_calc_merge_features}
\title{Merge input data.frame objects}
\usage{
post_calc_merge_features(by = c("site_id"), time = FALSE, ...)
}
\arguments{
\item{by}{character. Joining keys. See \code{\link{merge}} for details.}

\item{time}{logical(1). Whether or not include time identifier.
Set this \code{TRUE} will supersede \code{by} value by appending time identifier.}

\item{...}{data.frame objects to merge}
}
\value{
data.table
}
\description{
Merge input data.frame objects
}
\seealso{
Other Post-calculation: 
\code{\link{add_time_col}()},
\code{\link{impute_all}()},
\code{\link{post_calc_autojoin}()},
\code{\link{post_calc_convert_time}()},
\code{\link{post_calc_merge_all}()},
\code{\link{post_calc_unify_timecols}()},
\code{\link{reduce_merge}()}
}
\concept{Post-calculation}
