% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_base_functions.R
\name{post_calc_unify_timecols}
\alias{post_calc_unify_timecols}
\title{Change time column name}
\usage{
post_calc_unify_timecols(df, candidates = c("year"), replace = "time")
}
\arguments{
\item{df}{data.frame}

\item{candidates}{character. Candidate column names.}

\item{replace}{character. New column name.}
}
\value{
data.frame
}
\description{
Change time column name
}
\seealso{
Other Post-calculation: 
\code{\link{add_time_col}()},
\code{\link{append_predecessors}()},
\code{\link{impute_all}()},
\code{\link{post_calc_autojoin}()},
\code{\link{post_calc_convert_time}()},
\code{\link{post_calc_df_year_expand}()},
\code{\link{post_calc_join_yeardate}()},
\code{\link{post_calc_merge_all}()},
\code{\link{post_calc_merge_features}()},
\code{\link{post_calc_year_expand}()},
\code{\link{reduce_merge}()}
}
\concept{Post-calculation}
