% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_learner.R
\name{pred_colname}
\alias{pred_colname}
\title{Assign column name to base learner prediction based on hyperparameters.}
\usage{
pred_colname(pred)
}
\arguments{
\item{pred}{Base learner outcomes.}
}
\value{
a character for column name.
}
\description{
Assign column name to base learner prediction based on hyperparameters.
}
\keyword{Utility}
\keyword{meta_learner}
