% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_learner.R
\name{predict_meta_learner}
\alias{predict_meta_learner}
\title{Predict meta learner}
\usage{
predict_meta_learner(meta_fitted, new_data)
}
\arguments{
\item{meta_fitted}{Fitted meta learner model.}

\item{new_data}{data.frame. New data. Must have the same
predictands and predictors as the training data.}
}
\value{
Predicted values.
}
\description{
Predict meta learner
}
\keyword{meta_learner}
