% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_learner.R
\name{prepare_cvindex}
\alias{prepare_cvindex}
\title{Prepare spatial and spatiotemporal cross validation sets}
\usage{
prepare_cvindex(
  data,
  r_subsample = 0.3,
  target_cols = c("lon", "lat"),
  cv_make_fun = spatialsample::spatial_block_cv,
  ...
)
}
\arguments{
\item{data}{data.table with X, Y, and time information.}

\item{r_subsample}{The proportion of rows to be sampled.}

\item{target_cols}{character(3). Names of columns for X, Y.
Default is \code{c("lon", "lat")}. It is passed to sf::st_as_sf to
subsequently generate spatial cross-validation indices using
\code{spatialsample::spatial_block_cv} and
\code{spatialsample::spatial_clustering_cv}.}

\item{cv_make_fun}{function(1). Function to generate spatial
cross-validation indices. Default is \code{spatialsample::spatial_block_cv}.}
}
\value{
rsample::manual_rset() object.
}
\description{
Prepare spatial and spatiotemporal cross validation sets
}
\seealso{
\code{\link{generate_cv_index}} \code{\link[spatialsample:spatial_block_cv]{spatialsample::spatial_block_cv}}
\code{\link[spatialsample:spatial_clustering_cv]{spatialsample::spatial_clustering_cv}}
}
\keyword{Baselearner}
