% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_base_functions.R
\name{process_counties}
\alias{process_counties}
\title{Load county sf object}
\usage{
process_counties(
  year = 2020,
  exclude = c("02", "15", "60", "66", "68", "69", "72", "78")
)
}
\arguments{
\item{year}{integer(1). Year of the county shapefile.}

\item{exclude}{character. State FIPS codes to exclude.
Default is c("02", "15", "60", "66", "68", "69", "72", "78").}
}
\value{
sf object
}
\description{
Load county sf object
}
