% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_base_functions.R
\name{process_geos_bulk}
\alias{process_geos_bulk}
\title{Process atmospheric composition data by chunks (v2)}
\usage{
process_geos_bulk(path = NULL, date = c("2018-01-01", "2018-01-01"), ...)
}
\arguments{
\item{path}{character(1). Directory with downloaded netCDF (.nc4) files. or
netCDF file paths.}

\item{date}{character(2). length of 10. Format "YYYY-MM-DD".}

\item{...}{Arguments passed to \code{\link[terra:rast]{terra::rast}}.}
}
\value{
a \code{SpatRaster} object;
}
\description{
Returning a single \code{SpatRasterDataset} object.
}
\note{
Layer names of the returned \code{SpatRaster} object contain the variable,
pressure level, date
Reference duration: 1 day summary, all layers: 115 seconds
Superseded by \code{\link{calc_geos_strict}}.
}
\author{
Mitchell Manware, Insang Song
}
\keyword{Calculation}
