% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{query_modis_files}
\alias{query_modis_files}
\title{Identify MODIS files}
\usage{
query_modis_files(path, list, index)
}
\arguments{
\item{path}{A character vector specifying the path to the MODIS data.}

\item{list}{A list of julian dates.}

\item{index}{An integer specifying the index of the julian date to use.}
}
\value{
A character vector of MODIS file paths.
}
\description{
This function identifies the relevant MODIS file paths based on
path, list of julian dates, and index. Designed to help set arguments
for the \code{inject_modis_par} function.
}
\keyword{Calculation}
