% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_base_functions.R
\name{read_locs}
\alias{read_locs}
\title{Read AQS data}
\usage{
read_locs(fun_aqs = amadeus::process_aqs, export = FALSE, ...)
}
\arguments{
\item{fun_aqs}{function to import AQS data.
Default is \code{amadeus::process_aqs}}

\item{export}{Export the file to qs. Default is FALSE.}

\item{...}{Passed arguments to \code{fun_aqs}}
}
\value{
Depending on \code{fun_aqs} specification.
}
\description{
Read AQS data
}
\seealso{
Other Pipeline utility: 
\code{\link{attach_xy}()},
\code{\link{df_params}()},
\code{\link{feature_raw_download}()},
\code{\link{load_modis_files}()},
\code{\link{loadargs}()},
\code{\link{read_paths}()},
\code{\link{search_function}()},
\code{\link{set_slurm_resource}()}
}
\concept{Pipeline utility}
