% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_base_functions.R
\name{read_paths}
\alias{read_paths}
\title{Read paths from a directory with a specific file extension}
\usage{
read_paths(
  path,
  extension = ".hdf",
  target_dates = c("2020-01-01", "2020-01-15"),
  julian = FALSE
)
}
\arguments{
\item{path}{The directory path from which to read the paths.}

\item{extension}{The file extension to match. Defaults to ".hdf".}

\item{target_dates}{A character vector of length 2 containing
the start and end dates.}

\item{julian}{logical(1). If \code{TRUE}, the dates are in Julian format.}
}
\value{
A character vector containing the full paths of the matching files.
}
\description{
Read paths from a directory with a specific file extension
}
\examples{
\dontrun{
# Read paths from a directory with default extension
read_paths("/path/to/directory")

# Read paths from a directory with custom extension
read_paths("/path/to/directory", ".txt")
}
}
\seealso{
Other Pipeline utility: 
\code{\link{df_params}()},
\code{\link{feature_raw_download}()},
\code{\link{load_modis_files}()},
\code{\link{read_locs}()},
\code{\link{search_function}()},
\code{\link{set_slurm_resource}()}
}
\concept{Pipeline utility}
