% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{reduce_list}
\alias{reduce_list}
\title{Combine dynamically branched sublists based on common column names}
\usage{
reduce_list(df_list)
}
\arguments{
\item{df_list}{list(1). List of data frames with common column names.
Typically the output of a dynamically branched target.}
}
\value{
a list object, with data frames combined by common column names.
}
\description{
This function combines the sublists of dynamically branched targets object
as a preprocessor for \code{reduce_merge} function.
}
\keyword{prediction}
