% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/injection.R
\name{reduce_merge}
\alias{reduce_merge}
\title{Reduce and merge a list of data tables}
\usage{
reduce_merge(list_in, by = c("site_id", "time"), all.x = TRUE, all.y = FALSE)
}
\arguments{
\item{list_in}{A list of data tables to be merged.}

\item{by}{The columns to merge the data tables on.
If \code{NULL}, the function will automatically detect the common column names.}

\item{all.x}{logical(1). Keeping all rows from the first input.}

\item{all.y}{logical(1). Keeping all rows from the second input.}
}
\value{
A merged data table.
}
\description{
This function takes a list of data tables and merges them together
using the specified columns. It uses the \code{merge.data.table} function
from the \code{data.table} package to perform the merge.
}
\examples{
\dontrun{
# Create example data tables
dt1 <- data.table(a = 1:3, b = 4:6)
dt2 <- data.table(a = 2:4, c = 7:9)
dt3 <- data.table(a = 3:5, d = 10:12)

# Merge the data tables
reduce_merge(list(dt1, dt2, dt3), by = "a")
}
}
\keyword{Post-calculation}
