% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/injection.R
\name{reduce_merge_iter}
\alias{reduce_merge_iter}
\title{Reduce and merge a list of data tables}
\usage{
reduce_merge_iter(
  list_in,
  by = c("site_id", "time"),
  all.x = TRUE,
  all.y = FALSE
)
}
\arguments{
\item{list_in}{A list of data tables to be merged.}

\item{by}{The columns to merge the data tables on.
If \code{NULL}, the function will automatically detect the common column names.}

\item{all.x}{logical(1). Keeping all rows from the first input.}

\item{all.y}{logical(1). Keeping all rows from the second input.}
}
\value{
A merged data table.
}
\description{
This function iteratively runs \code{reduce_merge} on a list of \code{data.table}
objects for merging large \code{data.tables} that may not fit into memory
with \code{reduce_merge} directly.
}
\seealso{
\code{\link{reduce_merge}}
}
\keyword{Post-calculation}
