% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_base_functions.R
\name{restore_fit_best}
\alias{restore_fit_best}
\title{Restore the full data set from two rset objects then fit the best model}
\usage{
restore_fit_best(
  rset_trimmed,
  rset_full,
  df_full,
  by = c("site_id", "time"),
  nested = TRUE,
  nest_length = 30L
)
}
\arguments{
\item{rset_trimmed}{rset object without data in splits column.}

\item{rset_full}{rset object with full data.}

\item{df_full}{data.table with full data.}

\item{nested}{logical(1). If TRUE, the rset object is nested.}

\item{nest_length}{integer(1). Length of the nested list.
i.e., Number of resamples.}
}
\value{
rset object with full data in splits column.
}
\description{
Restore the full data set from two rset objects then fit the best model
}
\keyword{Baselearner}
