% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_base_functions.R
\name{restore_rset_full}
\alias{restore_rset_full}
\title{Restore the full data set from the rset object}
\usage{
restore_rset_full(rset, data_full)
}
\arguments{
\item{rset}{rsample::manual_rset() object's \code{splits} column}

\item{data_full}{data.table with all features}
}
\value{
A list of data.table objects.
}
\description{
Restore the full data set from the rset object
}
\note{
$splits should be present in rset.
}
\keyword{Baselearner}
