% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_base_functions.R
\name{search_function}
\alias{search_function}
\title{Search package functions}
\usage{
search_function(package, search)
}
\arguments{
\item{package}{character(1). Package name.}

\item{search}{character(1). Search term.}
}
\value{
A character vector containing the matching function names.
}
\description{
Search package functions
}
\examples{
# Search for functions in the `amadeus` package
search_function("amadeus", "process_")
}
