% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_base_functions.R
\name{set_args_calc}
\alias{set_args_calc}
\title{Set arguments for the calculation process}
\usage{
set_args_calc(
  char_siteid = "site_id",
  char_timeid = "time",
  char_period = c("2018-01-01", "2022-10-31"),
  num_extent = c(-126, -62, 22, 52),
  char_user_email = paste0(Sys.getenv("USER"), "@nih.gov"),
  export = FALSE,
  path_export = "inst/targets/punchcard_calc.qs",
  char_input_dir = "input",
  nthreads_nasa = 14L,
  nthreads_tri = 5L,
  nthreads_geoscf = 10L,
  nthreads_gmted = 4L,
  nthreads_narr = 24L,
  nthreads_groads = 3L,
  nthreads_population = 3L,
  nthreads_append = 8L,
  nthreads_impute = 64L
)
}
\arguments{
\item{char_siteid}{Character string specifying the site ID.
Default is "site_id".}

\item{char_timeid}{Character string specifying the time ID.
Default is "time".}

\item{char_period}{Character vector specifying the time period.
Default is c("2018-01-01", "2022-10-31").}

\item{num_extent}{Numeric vector specifying the extent.
Default is c(-126, -62, 22, 52).}

\item{char_user_email}{Character string specifying the user email.
Default is the current user's email with nih.gov domain.}

\item{export}{logical(1). If TRUE, the list for the calculation process
is exported to \code{path_export}. Default is FALSE.}

\item{path_export}{Character string specifying the export path.
Default is "inst/targets/punchcard_calc.qs".}

\item{char_input_dir}{Character string specifying the input path.
Default is "input".}

\item{nthreads_nasa}{integer(1). Number of threads for NASA data.
Default is 14L.}

\item{nthreads_tri}{integer(1). Number of threads for TRI data.
Default is 5L.}

\item{nthreads_geoscf}{integer(1). Number of threads for GEOSCF data.
Default is 10L.}

\item{nthreads_gmted}{integer(1). Number of threads for GMTED data.
Default is 4L.}

\item{nthreads_narr}{integer(1). Number of threads for NARR data.
Default is 24L.}

\item{nthreads_groads}{integer(1). Number of threads for GROADS data.
Default is 3L.}

\item{nthreads_population}{integer(1). Number of threads for population data.
Default is 3L.}

\item{nthreads_append}{integer(1). Number of threads for appending data.
Default is 8L.}

\item{nthreads_impute}{integer(1). Number of threads for imputing data.
Default is 64L.}
}
\value{
A list of arguments for common use
in the calculation process.
\itemize{
\item char_siteid: Character string specifying the site ID.
\item char_timeid: Character string specifying the time ID.
\item char_period: Character vector specifying the time period.
\item num_extent: Numeric vector specifying the extent.
\item char_user_email: Character string specifying the user email.
\item char_input_dir: Character string specifying the input path.
\item nthreads_nasa: Number of threads for NASA data.
\item nthreads_tri: Number of threads for TRI data.
\item nthreads_geoscf: Number of threads for GEOS-CF data.
\item nthreads_gmted: Number of threads for GMTED data.
\item nthreads_narr: Number of threads for NARR data.
\item nthreads_groads: Number of threads for SEDAC Groads data.
\item nthreads_population: Number of threads for population data.
\item nthreads_append: Number of threads for appending data.
\item nthreads_impute: Number of threads for imputing data.
}
}
\description{
This function sets the arguments for the calculation process.
It takes several parameters including site ID, time ID, time period,
extent, user email, export path, and input path.
}
\note{
The number of threads used is fixed as 1L
otherwise specified in \verb{nthreads_*} arguments.
path_input should contain the following subdirectories:
\itemize{
\item modis/raw/61/MOD11A1
\item modis/raw/61/MOD06_L2
\item modis/raw/61/MOD09GA
\item modis/raw/61/MCD19A2
\item modis/raw/61/MOD13A2
\item modis/raw/5000/VNP46A2
\item aqs
\item nlcd
\item geos/aqc_tavg_1hr_g1440x721_v1
\item geos/chm_tavg_1hr_g1440x721_v1
\item HMS_Smoke/data
\item gmted
\item nei
\item narr
\item HMS_Smoke
\item koppen_geiger
\item ecoregions
\item sedac_groads
\item sedac_population
}
}
\author{
Insang Song
}
\keyword{Utility}
