% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_base_functions.R
\name{set_args_download}
\alias{set_args_download}
\title{Generate argument list for raw data download}
\usage{
set_args_download(
  char_period = c("2018-01-01", "2022-10-31"),
  char_input_dir = "input",
  path_export = "inst/targets/punchcard_download.qs"
)
}
\description{
Generate argument list for raw data download
}
\keyword{Utility}
