% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{startup_nrtap}
\alias{startup_nrtap}
\title{Start-up settings}
\usage{
startup_nrtap(
  workdir = "./",
  user_mode = TRUE,
  username = Sys.info()["effective_user"]
)
}
\arguments{
\item{workdir}{character(1). Working directory path.}

\item{user_mode}{logical(1).
Use an user directory for fitting models.
When \code{TRUE}, a directory with the current session's user name will be
created under [working directory]/output/[username].}

\item{username}{character(1). Used username. Default is
the "effective_user" element of \code{[Sys.info()]} outputs.}
}
\description{
Start-up settings
}
