% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_learner.R
\name{switch_model}
\alias{switch_model}
\title{Define a base learner model based on parsnip and tune}
\usage{
switch_model(
  model_type = c("mlp", "mlp2", "xgb", "lgb", "elnet", "brulee_linear"),
  device = c("cpu", "cuda", "gpu")
)
}
\arguments{
\item{model_type}{character(1). Model type to be used.
Default is "mlp". Available options are "mlp", "xgb", "lgb", "elnet".}

\item{device}{character(1). The device to be used for training.
GPU acceleration is possible for \code{brulee}-engine multi-layer
perceptron (\code{model_type = "mlp"};\code{device = "cuda"}) and for \code{lightgbm}-engine
boosted tree (\verb{model_type = "lgb"; device = "gpu"}). Ensure your system is
equipped with CUDA-enabled graphical processing units if utilizing \code{cuda} or
\code{gpu}.}
}
\value{
A parsnip model object.
}
\description{
Define a base learner model based on parsnip and tune
}
\keyword{Baselearner}
