% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_arguments.R
\name{sys_beethoven}
\alias{sys_beethoven}
\title{Configure Library Paths and Environment Variables for Beethoven Workflow}
\usage{
sys_beethoven(
  libpaths = .libPaths(grep(paste0("biotools|", Sys.getenv("USER")), .libPaths(), value =
    TRUE, invert = TRUE)),
  path = paste0("/usr/local/cuda/bin:", "/usr/local/nvidia/bin:", "/usr/local/cuda/bin:",
    "/usr/local/sbin:", "/usr/local/bin:", "/usr/sbin:", "/usr/bin:", "/sbin:", "/bin"),
  ld_library_path = "/usr/local/cuda/lib64",
  cuda_home = "/usr/local/cuda"
)
}
\arguments{
\item{libpaths}{A character vector specifying the library paths to use.
By default, it excludes user-specific and host paths from \code{.libPaths()}.}

\item{path}{A character string specifying the system \code{PATH} environment
variable. Defaults to a container-friendly configuration with CUDA paths.}

\item{ld_library_path}{A character string specifying the \code{LD_LIBRARY_PATH}
environment variable. Defaults to \code{"/usr/local/cuda/lib64"}.}

\item{cuda_home}{A character string specifying the \code{CUDA_HOME} environment
variable. Defaults to \code{"/usr/local/cuda"}.}
}
\description{
This function sets up the library paths and environmental variables requireds_
for running the Beethoven workflow in a containerized environment.
}
\keyword{Utility}
