% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_learner.R
\name{vec2nc}
\alias{vec2nc}
\title{vec2nc takes a vector or matrix of predictions, some spatial information,
and create a NetCDF}
\usage{
vec2nc(pred_vec, pred_bbox, field_names)
}
\arguments{
\item{pred_vec}{vector/matrix of predictions}

\item{pred_bbox}{an sf bounding box defining the domain of the NetCDF}
}
\value{
nc_out a NetCDF file
}
\description{
vec2nc takes a vector or matrix of predictions, some spatial information,
and create a NetCDF
}
\examples{
NULL
}
