% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_base_functions.R
\name{vis_rset}
\alias{vis_rset}
\title{Visualize the spatio-temporal cross-validation index}
\usage{
vis_rset(rsplit)
}
\arguments{
\item{rsplit}{rsample::manual_rset() object.}
}
\value{
None. A plot will be generated.
}
\description{
Visualize the spatio-temporal cross-validation index
}
\seealso{
\code{\link{generate_cv_index}}

Other Base learner: 
\code{\link{convert_cv_index_rset}()}
}
\concept{Base learner}
