% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_learner.R
\name{vis_spt_rset}
\alias{vis_spt_rset}
\title{Visualize the spatio-temporal cross-validation index}
\usage{
vis_spt_rset(rsplit, cex = 0.02, angle = 60)
}
\arguments{
\item{rsplit}{rsample::manual_rset() object.}

\item{cex}{numeric(1). Size of the points in the plot.}

\item{angle}{numeric(1). Viewing angle of 3D plot.}
}
\value{
None. A plot will be generated.
}
\description{
Visualize the spatio-temporal cross-validation index
}
\keyword{Baselearner}
