% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_learner.R
\name{attach_pred}
\alias{attach_pred}
\title{Combine prediction values from base learners}
\usage{
attach_pred(
  data,
  pred,
  position = 1,
  target_cols = c("site_id", "time", "Event.Type", "lon", "lat"),
  yvar = "Arithmetic.Mean"
)
}
\arguments{
\item{data}{data.frame(1). Full data.}

\item{pred}{list(1). List with base learner prediction values.}

\item{position}{numeric(1). Position of the prediction values in the list
\code{pred}. Default is 1.}

\item{target_cols}{characters(1). Columns to retain from the full
data.frame.}

\item{yvar}{character(1). Outcome variable name.}
}
\value{
a data.frame object, including the target columns from \code{data} and
the predictions for each base learner.
}
\description{
This function combines outcome data (observations) with the prediction
values from each base learner.
}
\keyword{Utility}
\keyword{meta_learner}
