% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_learner.R
\name{base_learner_fit_cnn}
\alias{base_learner_fit_cnn}
\title{Fit convolutional neural networks with neural network library}
\usage{
base_learner_fit_cnn(ymat, xmat)
}
\arguments{
\item{ymat}{torch::torch_tensor. Dependent variable.}

\item{xmat}{torch::torch_tensor. Independent variables.}
}
\value{
data.frame
}
\description{
This function uses torch as a backend.
Torch currently supports CPU, CUDA, and Metal (Apple Silicon graphics),
thus users can leverage available computational assets
to accelerate the model fitting.
}
\author{
Insang Song
}
