% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_base_functions.R
\name{calc_geos_strict}
\alias{calc_geos_strict}
\title{Process atmospheric composition data by chunks (v3)}
\usage{
calc_geos_strict(
  path = NULL,
  date = c("2018-01-01", "2018-01-01"),
  locs = NULL,
  locs_id = NULL,
  ...
)
}
\arguments{
\item{path}{character(1). Directory with downloaded netCDF (.nc4) files. or
netCDF file paths.}

\item{date}{character(2). length of 10. Format "YYYY-MM-DD".}

\item{...}{Arguments passed to \code{\link[terra:rast]{terra::rast}}.}
}
\value{
a \code{SpatRaster} object;
}
\description{
Returning a single \code{SpatRasterDataset} object.
Removed \code{tapp} for performance; impose a strict assumption that
there are no missing values
}
\note{
Layer names of the returned \code{SpatRaster} object contain the variable,
pressure level, date
Reference duration: 1 day summary, all layers: 106 seconds
hard-coded subsets for subdataset selection
}
\author{
Mitchell Manware, Insang Song
}
