% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_base_functions.R
\name{calc_gmted_direct}
\alias{calc_gmted_direct}
\title{Reflown gmted processing}
\usage{
calc_gmted_direct(
  variable = NULL,
  path = NULL,
  locs = NULL,
  locs_id = NULL,
  win = c(-126, -62, 22, 52),
  radius = 0,
  fun = "mean",
  ...
)
}
\arguments{
\item{variable}{character(2). Statistic and resolution.}

\item{path}{character(1). Directory with downloaded GMTED files.}

\item{locs}{data.frame/SpatVector/sf. Locations.}

\item{locs_id}{character(1). Location identifier.}

\item{win}{numeric(4). Window for the raster.}

\item{radius}{numeric(1). Radius for the extraction.}

\item{fun}{character(1). Function to apply.}

\item{...}{Additional parameters to be passed to other functions.}
}
\value{
A data.frame containing the extracted GMTED data.
}
\description{
Reflown gmted processing
}
\seealso{
Other Calculation: 
\code{\link{calc_geos_strict}()},
\code{\link{calc_narr2}()},
\code{\link{calculate}()},
\code{\link{inject_calculate}()},
\code{\link{inject_geos}()},
\code{\link{inject_gmted}()},
\code{\link{inject_modis_par}()},
\code{\link{par_narr}()},
\code{\link{process_counties}()},
\code{\link{process_geos_bulk}()},
\code{\link{process_narr2}()}
}
\concept{Calculation}
