% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_base_functions.R
\name{calc_narr2}
\alias{calc_narr2}
\title{Calculate aggregated values for specified locations}
\usage{
calc_narr2(from, locs, locs_id = NULL, radius = 0, fun = "mean", ...)
}
\arguments{
\item{from}{The raster dataset from which to extract values.}

\item{locs}{A data frame containing the locations for which to calculate aggregated values.
It should have a column named "site_id" that contains unique identifiers for each location.}

\item{locs_id}{An optional column name in the \code{locs} data frame that contains additional location identifiers.}

\item{radius}{The radius within which to include neighboring locations for aggregation. Default is 0.}

\item{fun}{The aggregation function to use. It can be a character string specifying a function name (e.g., "mean", "sum"),
or it can be a custom function. Default is "mean".}

\item{...}{Additional arguments to be passed to the aggregation function.}
}
\value{
A data frame containing the aggregated values for each location and time point.
}
\description{
This function calculates aggregated values for specified locations from a raster dataset.
}
\examples{
# Calculate mean values for locations within a radius of 10 units
calc_narr2(from = raster_data, locs = locations, radius = 10, fun = "mean")

}
