% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calc_tri_mod}
\alias{calc_tri_mod}
\title{Modified isotropic Sum of Exponentially Decaying Contributions (SEDC)
covariates}
\usage{
calc_tri_mod(from = NULL, locs, locs_id = "site_id", radius = 1000L, ...)
}
\arguments{
\item{from}{SpatVector(1). Output of \code{process_tri()}.}

\item{locs}{sf/SpatVector. Locations where TRI variables are calculated.}

\item{locs_id}{character(1). Unique site identifier column name.
Default is \code{"site_id"}.}

\item{radius}{integer(1). Circular buffer radius.}

\item{...}{Placeholders.}
}
\value{
a data.frame or SpatVector object
}
\description{
Calculate toxic release values for polygons or isotropic buffer point
locations. Returns a \code{data.frame} object containing \code{locs_id}
and variables for each chemical in \code{from}.
}
\note{
U.S. context.
}
\author{
Insang Song, Mariana Kassien
}
\keyword{Calculation}
