% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_base_functions.R
\name{calculate}
\alias{calculate}
\title{Spatiotemporal covariate calculation}
\usage{
calculate(
  domain = NULL,
  domain_name = "year",
  nthreads = 1L,
  process_function = amadeus::process_covariates,
  calc_function = amadeus::calc_covariates,
  ...
)
}
\arguments{
\item{domain}{vector of integer/character/Date.
Depending on temporal resolution of raw datasets.
Nullable; If \code{NULL}, it will be set to \code{c(1)}.}

\item{domain_name}{character(1). Name of the domain. Default is \code{"year"}.}

\item{process_function}{Raw data processor. Default is
\code{\link[amadeus:process_covariates]{amadeus::process_covariates}}}

\item{calc_function}{Function to calculate covariates.
\code{\link[amadeus:calc_covariates]{amadeus::calc_covariates}}}

\item{...}{Arguments passed to \code{process_function} and \code{calc_function}}
}
\value{
A data.table object.
}
\description{
Spatiotemporal covariate calculation
}
\examples{
df_fine0 <- data.frame(site_id = c("A", "B", "B", "C"),
                      lon = rep("barns", 4),
                      time = as.Date(c("2022-01-01", "2022-01-02", "2021-12-31", "2021-01-03")),
                      value = c(1, 2, 3, 5))
df_coarse0 <- data.frame(site_id = c("A", "B", "C"),
                       lon = rep("J", 3),
                        time = c("2022", "2022", "2021"),
                        other_value = c(10, 20, 30))
jdf <- post_calc_autojoin(df_fine0, df_coarse0)
print(jdf)
}
