% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_base_functions.R
\name{convert_cv_index_rset}
\alias{convert_cv_index_rset}
\title{Generate manual rset object from spatiotemporal cross-validation indices}
\usage{
convert_cv_index_rset(cvindex, data, ref_list = NULL, cv_mode = "spt")
}
\arguments{
\item{cvindex}{integer length of nrow(data).}

\item{data}{data.frame.}

\item{ref_list}{List of custom reference indices. Default is NULL.
if not NULL, it will be used as a reference instead of max(cvindex).}

\item{cv_mode}{character(1). Spatiotemporal cross-validation indexing
method label.}
}
\value{
rset object of \code{rsample} package. A tibble with a list column of
training-test data.frames and a column of labels.
}
\description{
Generate manual rset object from spatiotemporal cross-validation indices
}
\author{
Insang Song
}
\keyword{Base}
\keyword{learner}
