% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_base_functions.R
\name{df_params}
\alias{df_params}
\title{Get data.frame of function parameters}
\usage{
df_params(functions)
}
\arguments{
\item{functions}{character. Vector of function names.}
}
\value{
A data.frame containing the parameters of the functions.
}
\description{
Get data.frame of function parameters
}
