% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_base_functions.R
\name{divisor}
\alias{divisor}
\title{Get Divisors}
\usage{
divisor(x)
}
\arguments{
\item{x}{integer(1). A positive integer.}
}
\value{
A vector of divisors of x.
}
\description{
Get Divisors
}
\seealso{
Other Miscellaneous: 
\code{\link{\%tin\%}()}
}
\concept{Miscellaneous}
