% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_base_functions.R
\name{feature_raw_download}
\alias{feature_raw_download}
\title{Check file status and download if necessary}
\usage{
feature_raw_download(path = NULL, dataset_name = NULL, ...)
}
\arguments{
\item{path}{download path.}

\item{dataset_name}{Dataset name. See \code{\link[amadeus:download_data]{amadeus::download_data}} for details.}

\item{...}{Arguments passed to \code{amadeus::download_data}}
}
\value{
logical(1).
}
\description{
Check file status and download if necessary
}
\seealso{
Other Pipeline utility: 
\code{\link{attach_xy}()},
\code{\link{df_params}()},
\code{\link{load_modis_files}()},
\code{\link{loadargs}()},
\code{\link{read_locs}()},
\code{\link{read_paths}()},
\code{\link{search_function}()},
\code{\link{set_slurm_resource}()}
}
\concept{Pipeline utility}
