% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_base_functions.R
\name{fit_base_brulee}
\alias{fit_base_brulee}
\title{Base learner: Multilayer perceptron with brulee}
\usage{
fit_base_brulee(
  dt_imputed,
  folds = NULL,
  r_subsample = 0.3,
  yvar = "Arithmetic.Mean",
  xvar = seq(6, ncol(dt_imputed)),
  vfold = 5L,
  ...
)
}
\arguments{
\item{dt_imputed}{The input data table to be used for fitting.}

\item{folds}{pre-generated rset object. If NULL, it should be
numeric to be used in \link[rsample:vfold_cv]{rsample::vfold_cv}.}

\item{r_subsample}{The proportion of rows to be sampled.}

\item{yvar}{The target variable.}

\item{xvar}{The predictor variables.}

\item{vfold}{The number of folds for cross-validation.}

\item{...}{Additional arguments to be passed.}
}
\value{
The fitted workflow.
}
\description{
Multilayer perceptron model with different configurations of
hidden units, dropout, activation, and learning rate using brulee
and tidymodels. With proper settings, users can utilize graphics
processing units (GPU) to speed up the training process.
}
\note{
Spatiotemporal cross-validation strategy is not yet implemented.
tune package should be 1.2.0 or higher.
}
\keyword{Base}
\keyword{learner}
