% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_base_functions.R
\name{fit_base_elnet}
\alias{fit_base_elnet}
\title{Base learner: Elastic net}
\usage{
fit_base_elnet(
  dt_imputed,
  folds = NULL,
  r_subsample = 0.3,
  yvar = "Arithmetic.Mean",
  xvar = seq(6, ncol(dt_imputed)),
  vfold = 5L,
  nthreads = 16L,
  ...
)
}
\arguments{
\item{dt_imputed}{The input data table to be used for fitting.}

\item{folds}{pre-generated rset object. If NULL, it should be
numeric to be used in \link[rsample:vfold_cv]{rsample::vfold_cv}.}

\item{r_subsample}{The proportion of rows to be sampled.}

\item{yvar}{The target variable.}

\item{xvar}{The predictor variables.}

\item{vfold}{The number of folds for cross-validation.}

\item{nthreads}{The number of threads to be used. Default is 16L.}

\item{...}{Additional arguments to be passed.}
}
\value{
The fitted workflow.
}
\description{
Elastic net model is fitted at the defined rate (\code{r_subsample}) of
the input dataset by grid search.
}
\note{
Spatiotemporal cross-validation strategy is not yet implemented.
}
\keyword{Base}
\keyword{learner}
