% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_base_functions.R
\name{fit_base_xgb}
\alias{fit_base_xgb}
\title{Base learner: Extreme gradient boosting (XGBoost)}
\usage{
fit_base_xgb(
  dt_imputed,
  r_subsample = 0.3,
  yvar = "Arithmetic.Mean",
  xvar = seq(6, ncol(dt_imputed)),
  vfold = 5L,
  ...
)
}
\arguments{
\item{dt_imputed}{The input data table to be used for fitting.}

\item{r_subsample}{The proportion of rows to be sampled.}

\item{yvar}{The target variable.}

\item{xvar}{The predictor variables.}

\item{vfold}{The number of folds for cross-validation.}

\item{...}{Additional arguments to be passed.}

\item{folds}{pre-generated rset object. If NULL, it should be
numeric to be used in \link[rsample:vfold_cv]{rsample::vfold_cv}.}
}
\value{
The fitted workflow.
}
\description{
XGBoost model is fitted at the defined rate (\code{r_subsample}) of
the input dataset by grid search.
With proper settings, users can utilize graphics
processing units (GPU) to speed up the training process.
}
\note{
Spatiotemporal cross-validation strategy is not yet implemented.
}
\seealso{
Other Base Learner: 
\code{\link{fit_base_brulee}()},
\code{\link{fit_base_elnet}()},
\code{\link{generate_cv_index}()}
}
\concept{Base Learner}
