% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_learner.R
\name{fit_meta_learner}
\alias{fit_meta_learner}
\title{Fit meta learner}
\usage{
fit_meta_learner(
  data,
  c_subsample = 0.5,
  r_subssample = 1,
  yvar = "Arithmetic.Mean",
  target_cols = c("site_id", "time", "Event.Type", "lon", "lat"),
  args_generate_cv = list(),
  tune_iter = 50L,
  nthreads = 2L,
  trim_resamples = FALSE,
  return_best = TRUE,
  metric = "rmse"
)
}
\arguments{
\item{data}{data.frame. Full data.frame of base learner predictions and AQS
spatiotemporal identifiers. \code{\link{attach_pred}}.}

\item{c_subsample}{numeric(1). Rate of column resampling. Default is 0.5.}

\item{yvar}{character(1). Outcome variable name}

\item{target_cols}{characters(1). Columns from \code{data} to be retained during
column resampling. Default is c("site_id", "time", "Event.Type", "lon",
"lat").}

\item{args_generate_cv}{List of arguments to be passed to
\code{switch_generate_cv_rset} function.}

\item{tune_iter}{integer(1). Bayesian optimization iterations.
Default is 50.}

\item{nthreads}{integer(1). The number of threads to be used for
tuning. Default is 8L. \code{learner = "elnet"} will utilize the multiple
threads in \code{\link[future:multicore]{future::multicore()}} plan. Passed to \code{\link{fit_base_tune}}.}

\item{trim_resamples}{logical(1). Default is TRUE, which replaces the actual
data.frames in splits column of \code{tune_results} object with NA.
Passed to \code{\link{fit_base_tune}}.}

\item{return_best}{logical(1). If TRUE, the best tuned model is returned.
Passed to \code{\link{fit_base_tune}}.}

\item{metric}{character(1). The metric to be used for selecting the best.
Must be one of "rmse", "rsq", "mae". Default = "rmse".
Passed to \code{\link{fit_base_tune}}.}

\item{r_subsample}{numeric(1). The proportion of rows to be used. Default is
1.0, which uses full dataset but setting is required to balance groups
generated with \code{\link{make_subdata}}}
}
\value{
List of 3, including the best-fit model, the best hyperparameters,
and the all performance records from \code{tune::tune_bayes()}.
Note that the meta learner function returns the best-fit model,
not predicted values.
}
\description{
This function subsets the full data by column subsamples (rate=50\%)
The optimal hyperparameter search is performed based on spatiotemporal
cross-validation schemes. As of version 0.4.5, users can define metric
used for selecting best hyperparameter set (default = "rmse").
}
\seealso{
\code{\link{fit_base_tune}} \code{\link{make_subdata}}
}
\keyword{meta_learner}
