% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{fl_dates}
\alias{fl_dates}
\title{Extract the first and last elements of a list}
\usage{
fl_dates(dates)
}
\arguments{
\item{dates}{list. A list of dates.}
}
\value{
a character vector with the first and last dates from the list.
}
\description{
Extract the first and last elements of a list
}
