% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{generate_cv_index}
\alias{generate_cv_index}
\title{Generate spatio-temporal cross-validation index}
\usage{
generate_cv_index(
  covars,
  cv_mode = c("lolo", "loto", "lolto", "random", "lblo", "lbto", "lblto"),
  cv_fold = 5L,
  sp_fold = NULL,
  t_fold = NULL,
  blocks = NULL,
  block_id = NULL
)
}
\arguments{
\item{covars}{stdt. See \link[NRTAPModel]{convert_stobj_to_stdt}
for details.}

\item{cv_mode}{character(1). One of 'lolo (leave-one-location-out)',
'loto (leave-one-time-out)',
'lolto (leave-one-location-time-out)',
'lblo (leave-block-location-out)',
'lbto (leave-block-time-out)',
'lblto (leave-block-location-time-out; not implemented)'
'random (full random selection)'}

\item{cv_fold}{integer(1). Number of folds for cross-validation.}

\item{sp_fold}{integer(1). Number of subfolds for spatial blocks.}

\item{t_fold}{integer(1). Number of subfolds for temporal blocks.}

\item{blocks}{integer(2)/sf/SpatVector object.}

\item{block_id}{character(1). The unique identifier of each block.}
}
\value{
A numeric vector with length of the number of input rows
}
\description{
Generate spatio-temporal cross-validation index
}
\details{
\code{blocks} is NULL as default
(i.e., no block structure is considered); then "lb*" cv_mode
is not working. If any meaningful object is assigned here,
cv_fold will not make sense and be ignored.
Definition of unit grid (horizontal, vertical) size,
arbitrary shape of blocks (sf/SpatVector case)
where users can consider common spatial hierarchies such as
states, counties, hydrological unit code areas, etc.
\code{lblto} mode will accept \code{sp_fold} and \code{t_fold}.
The maximum of results in this case is \code{sp_fold * t_fold}.
}
\author{
Insang Song
}
