% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_base_functions.R
\name{generate_cv_index}
\alias{generate_cv_index}
\title{Generate spatio-temporal cross-validation index with anticlust}
\usage{
generate_cv_index(
  data,
  target_cols = c("lon", "lat", "time"),
  cv_fold = 5L,
  cv_pairs = NULL,
  cv_mode = "spt"
)
}
\arguments{
\item{data}{data.table with X, Y, and time information.}

\item{target_cols}{character(3). Names of columns for X, Y, and time.
Default is c("lon", "lat", "time").
Order insensitive.}

\item{cv_fold}{integer(1). Number of folds for cross-validation.
default is 5L.}

\item{cv_pairs}{integer(1). Number of pairs for cross-validation.
This value will be used to generate a rank-based pairs
based on \code{target_cols} values.}

\item{cv_mode}{character(1). Spatiotemporal cross-validation indexing}
}
\value{
rsample::manual_rset() object.
}
\description{
This function generates a spatio-temporal cross-validation index
based on the anticlust package. The function first calculates the
spatial clustering index using the balanced_clustering function as
default, and if \code{cv_pairs} is provided,
it generates rank-based pairs based on the proximity between
cluster centroids.
}
\note{
nrow(data) \%\% cv_fold should be 0.
}
\seealso{
Other Base Learner: 
\code{\link{fit_base_brulee}()},
\code{\link{fit_base_elnet}()},
\code{\link{fit_base_xgb}()}
}
\author{
Insang Song
}
\concept{Base Learner}
