% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_base_functions.R
\name{generate_cv_index}
\alias{generate_cv_index}
\title{Generate spatio-temporal cross-validation index with anticlust}
\usage{
generate_cv_index(
  data,
  target_cols = c("lon", "lat", "time"),
  preprocessing = c("none", "normalize", "standardize"),
  cv_fold = 5L,
  cv_pairs = NULL,
  pairing = c("1", "2"),
  cv_mode = "spt"
)
}
\arguments{
\item{data}{data.table with X, Y, and time information.}

\item{target_cols}{character(3). Names of columns for X, Y, and time.
Default is c("lon", "lat", "time").
Order insensitive.}

\item{preprocessing}{character(1). Preprocessing method.
\itemize{
\item "none": no preprocessing.
\item "normalize": normalize the data.
\item "standardize": standardize the data.
}}

\item{cv_fold}{integer(1). Number of folds for cross-validation.
default is 5L.}

\item{cv_pairs}{integer(1). Number of pairs for cross-validation.
This value will be used to generate a rank-based pairs
based on \code{target_cols} values.}

\item{pairing}{character(1) Pair selection method.
\itemize{
\item "1": search the nearest for each cluster then others
are selected based on the rank.
\item "2": rank the pairwise distances directly
}}

\item{cv_mode}{character(1). Spatiotemporal cross-validation indexing}
}
\value{
rsample::manual_rset() object.
}
\description{
This function generates a spatio-temporal cross-validation index
based on the anticlust package. The function first calculates the
spatial clustering index using the balanced_clustering function as
default, and if \code{cv_pairs} is provided,
it generates rank-based pairs based on the proximity between
cluster centroids.
}
\note{
nrow(data) \%\% cv_fold should be 0.
}
\author{
Insang Song
}
\keyword{Base}
\keyword{learner}
