% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_learner.R
\name{generate_cv_index_spt}
\alias{generate_cv_index_spt}
\title{Generate spatio-temporal cross-validation index with
\code{spatialsample::spatial_block_cv} and year-based temporal folds}
\usage{
generate_cv_index_spt(
  data,
  crs = NULL,
  cellsize = NULL,
  locs_id = "site_id",
  coords = c("lon", "lat"),
  v = 10L,
  time_id = "time"
)
}
\arguments{
\item{data}{data.table or data.frame with \code{id}, \code{coords}, and \code{time} columns.}

\item{crs}{The coordinate reference system (CRS) of the spatial object}

\item{cellsize}{The size of each spatial block in meters. Uses the crs units}

\item{locs_id}{The column name in \code{data} that represents the
location identifier.}

\item{coords}{The column names in the spatial object that represent the
XY coordinates. Default is \code{c("lon", "lat")}.}

\item{v}{integer(1). The number of partitions for the resampling.}

\item{time_id}{The column name in \code{data} that represents the time values.}
}
\description{
This function generates spatio-temporal cross-validation indices with
\code{v} spatial blocks and year-based temporal folds. The spatial blocks
are generated with \code{spatialsample::spatial_block_cv} function, and the
temporal folds are generated based on the years availble in \code{data$time}.
Total number of folds is equal to
\code{v * length(unique(substr(data$time, 1, 4)))}.
}
\seealso{
\code{\link[spatialsample:spatial_block_cv]{spatialsample::spatial_block_cv}}
}
\keyword{Baselearner}
