% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_pipeline_example.R
\name{generic_base_learner}
\alias{generic_base_learner}
\title{define a base learner dummy: this illustrates one way to wrap a learning
algorithm and return an object that makes predictions}
\usage{
generic_base_learner(response, covariate, obs_locs, model_attr)
}
\arguments{
\item{response}{the dependent variable, numeric, vector}

\item{covariate}{independent variables, design matrix, numeric}

\item{obs_locs}{geospatial locations as sf object}

\item{model_attr}{option model attributes or metadata}
}
\description{
define a base learner dummy: this illustrates one way to wrap a learning
algorithm and return an object that makes predictions
}
