% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_base_functions.R
\name{inject_calculate}
\alias{inject_calculate}
\title{Injects the calculate function with specified arguments.}
\usage{
inject_calculate(covariate, locs, injection)
}
\arguments{
\item{covariate}{character(1). The name of the covariate to be calculated.}

\item{locs}{The locations to be used in the calculation.}

\item{injection}{Additional arguments to be injected into
the calculate function.}

\item{buffer}{The buffer size for the calculation. If not provided, the
default buffer size will be used.}
}
\value{
The result of the calculate function with the injected arguments.
}
\description{
This function injects the calculate function with the specified arguments,
allowing for dynamic customization of the function's behavior.
}
\examples{
\dontrun{
inject_calculate(
  locs = my_locs, buffer = 10, domain = my_domain,
  injection = list(arg1 = "value1", arg2 = "value2")
)
}
}
