% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_base_functions.R
\name{inject_calculate}
\alias{inject_calculate}
\title{Injects the calculate function with specified arguments.}
\usage{
inject_calculate(covariate, locs, injection)
}
\arguments{
\item{covariate}{character(1). The name of the covariate to be calculated.}

\item{locs}{The locations to be used in the calculation.}

\item{injection}{Additional arguments to be injected into
the calculate function.}
}
\value{
The result of the calculate function with the injected arguments.
}
\description{
This function injects the calculate function with the specified arguments,
allowing for dynamic customization of the function's behavior.
}
\examples{
\dontrun{
inject_calculate(
  locs = my_locs, buffer = 10, domain = my_domain,
  injection = list(arg1 = "value1", arg2 = "value2")
)
}
}
\seealso{
Other Calculation: 
\code{\link{calc_geos_strict}()},
\code{\link{calc_gmted_direct}()},
\code{\link{calc_narr2}()},
\code{\link{calculate}()},
\code{\link{inject_geos}()},
\code{\link{inject_gmted}()},
\code{\link{inject_modis_par}()},
\code{\link{par_narr}()},
\code{\link{process_counties}()},
\code{\link{process_geos_bulk}()},
\code{\link{process_narr2}()}
}
\concept{Calculation}
