% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_base_functions.R
\name{inject_modis_par}
\alias{inject_modis_par}
\title{Injects arguments to parallelize MODIS/VIIRS data processing}
\usage{
inject_modis_par(locs, domain, injection)
}
\arguments{
\item{locs}{A data frame containing the locations for which MODIS
features need to be calculated.}

\item{domain}{The domain in which the MODIS PAR data should be injected.}

\item{injection}{Additional parameters to be passed to the
\code{calc_modis_par} function.}
}
\value{
The modified domain with the injected MODIS PAR data.
}
\description{
Injects arguments to parallelize MODIS/VIIRS data processing
}
\seealso{
Other Calculation: 
\code{\link{calc_geos_strict}()},
\code{\link{calc_gmted_direct}()},
\code{\link{calc_narr2}()},
\code{\link{calculate}()},
\code{\link{inject_calculate}()},
\code{\link{inject_geos}()},
\code{\link{inject_gmted}()},
\code{\link{par_narr}()},
\code{\link{process_counties}()},
\code{\link{process_geos_bulk}()},
\code{\link{process_narr2}()}
}
\concept{Calculation}
