% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_base_functions.R
\name{loadargs}
\alias{loadargs}
\title{Load arguments from the formatted argument list file}
\usage{
loadargs(argfile, dataset)
}
\arguments{
\item{argfile}{character(1). Path to the argument file. RDS format.}

\item{dataset}{character(1). Dataset name.}
}
\value{
A list of arguments.
}
\description{
Load arguments from the formatted argument list file
}
\keyword{Pipeline}
\keyword{utility}
