% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_learner.R
\name{meta_learner_predict}
\alias{meta_learner_predict}
\title{Create meta_learner predictions from the list of BART fit objects and predictions of base learners}
\usage{
meta_learner_predict(meta_fit, base_outputs_stdt, nthreads = 2)
}
\arguments{
\item{meta_fit}{list of BART objects from meta_learner_fit}

\item{base_outputs_stdt}{stdt object. See \link{convert_stobj_to_stdt}
list with datatable containing lat, lon, time and the covariates
(outputs of each base learner) at prediction locations and crs.}

\item{nthreads}{integer(1). Number of threads used in \link[BART:predict.wbart]{BART::predict.wbart}}
}
\value{
meta_pred: the final meta learner predictions
}
\description{
The meta learner used in this package, Bayesian Additive Regression Tree (BART), is not explicitly a spatiotemporal model, but the input covariates (outputs of each base learner) are S-T based.
}
\note{
The predictions can be a rast or sf, which depends on the same
respective format of the covariance matrix input - cov_pred
}
\examples{
NULL
}
\references{
https://rspatial.github.io/terra/reference/predict.html
}
