% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_covariates.R
\name{modis_worker}
\alias{modis_worker}
\title{A single-date MODIS worker for parallelization}
\usage{
modis_worker(
  raster,
  date,
  sites_in = NULL,
  name_extracted = NULL,
  product = c("MOD11A1", "MOD13A2", "MOD06_L2", "VNP46A2", "MOD09GA", "MCD19A2"),
  fun_summary_raster = "mean",
  foo = scomps::extract_with_buffer,
  ...
)
}
\arguments{
\item{raster}{SpatRaster.}

\item{date}{Date(1). date to query.}

\item{sites_in}{sf object. AQS sites.}

\item{name_extracted}{character. Names of calculated covariates.}

\item{product}{character(1). Product code of MODIS. Should be one of
\code{c('MOD11A1', 'MOD13A2', 'MOD06_L2', 'VNP46A2', 'MOD09GA', 'MCD19A2')}}

\item{fun_summary_raster}{function. Summary function for
multilayer rasters. Passed to \code{foo}. See also
\code{\link[exactextractr]{exact_extract}} and
\code{\link[scomps]{extract_with}}.}

\item{foo}{function. A calculation function working with
SpatRaster and sf.}

\item{...}{Additional arguments passed to \code{foo}.}
}
\value{
A data.frame object.
}
\description{
A single-date MODIS worker for parallelization
}
\author{
Insang Song
}
