% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_base_functions.R
\name{post_calc_convert_time}
\alias{post_calc_convert_time}
\title{Convert time column to character}
\usage{
post_calc_convert_time(df)
}
\arguments{
\item{df}{data.table}
}
\description{
Convert time column to character
}
\note{
This function takes preprocessed data.table with
a column named \code{"time"}.
}
