% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_base_functions.R
\name{post_calc_drop_cols}
\alias{post_calc_drop_cols}
\title{Remove columns from a data frame based on regular expression patterns.}
\usage{
post_calc_drop_cols(
  df,
  candidates = "(^lon$|^lat$|geoid|year$|description|geometry)",
  strict = FALSE
)
}
\arguments{
\item{df}{The input data frame.}

\item{candidates}{A character vector of regular expression patterns
to match against column names. Columns that match any of the patterns
will be removed. The default value is
"^lon$|^lat$|geoid|year$|description".}

\item{strict}{logical(1). If \code{TRUE},
only \code{c("site_id", "time")} will be kept.}
}
\value{
The modified data frame with the specified columns removed.
}
\description{
Remove columns from a data frame based on regular expression patterns.
}
\examples{
\dontrun{
df <- data.frame(lon = 1:5, lat = 6:10, geoid = 11:15, year = 2010:2014,
                 description = letters[1:5], other = 16:20)
post_calc_drop_cols(df)
}
}
\keyword{"^lon$|^lat$|geoid|year$|description".}
\keyword{By}
\keyword{Post-calculation}
\keyword{This}
\keyword{a}
\keyword{any}
\keyword{columns}
\keyword{data}
\keyword{default,}
\keyword{expression}
\keyword{frame}
\keyword{from}
\keyword{function}
\keyword{it}
\keyword{match}
\keyword{names}
\keyword{of}
\keyword{patterns}
\keyword{patterns.}
\keyword{regular}
\keyword{removes}
\keyword{specified}
\keyword{that}
\keyword{the}
\keyword{with}
