% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_base_functions.R
\name{post_calc_join_yeardate}
\alias{post_calc_join_yeardate}
\title{Join a data.frame with a year-only date column to
that with a full date column}
\usage{
post_calc_join_yeardate(
  df_year,
  df_date,
  field_year = "time",
  field_date = "time",
  spid = "site_id"
)
}
\arguments{
\item{df_year}{data.frame with a year-only date column}

\item{df_date}{data.frame with a full date column}

\item{field_year}{character(1). Year column in \code{df_year}}

\item{field_date}{character(1). Date column in \code{df_date}}

\item{spid}{character(1). Name of the unique location identifier field.}
}
\value{
data.frame
}
\description{
The full date column will be converted to a year column
as a new column, then the data.frame with the year-only column will
be joined.
}
