% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_base_functions.R
\name{post_calc_merge_all}
\alias{post_calc_merge_all}
\title{Merge spatial and spatiotemporal covariate data}
\usage{
post_calc_merge_all(
  locs,
  locs_id,
  time_id,
  target_years = seq(2018, 2022),
  df_sp,
  df_spt
)
}
\arguments{
\item{locs}{Location. e.g., AQS sites.}

\item{locs_id}{character(1). Location identifier.}

\item{time_id}{character(1). Location identifier.}

\item{target_years}{integer. Used to dummify nominal year.}

\item{df_sp}{data.frame. Spatial-only covariates.}

\item{df_spt}{data.frame. Spatiotemporal covariates.}
}
\value{
data.frame
}
\description{
Merge spatial and spatiotemporal covariate data
}
\note{
This version assumes the time_id contains Date-like strings.
}
