% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_postprocessing.R
\name{post_calc_merge_features}
\alias{post_calc_merge_features}
\title{Merge input data.frame objects}
\usage{
post_calc_merge_features(by = c("site_id"), time = FALSE, ...)
}
\arguments{
\item{by}{character. Joining keys. See \code{\link{merge}} for details.}

\item{time}{logical(1). Whether or not include time identifier.
Set this \code{TRUE} will supersede \code{by} value by appending time identifier.}

\item{...}{data.frame objects to merge}
}
\value{
data.table
}
\description{
Merge input data.frame objects
}
\keyword{Post-calculation}
