% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_base_functions.R
\name{post_calc_unify_timecols}
\alias{post_calc_unify_timecols}
\title{Change time column name}
\usage{
post_calc_unify_timecols(df, candidates = c("year"), replace = "time")
}
\arguments{
\item{df}{data.frame}

\item{candidates}{character. Candidate column names.}

\item{replace}{character. New column name.}
}
\value{
data.frame
}
\description{
Change time column name
}
\keyword{Post-calculation}
