% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{process_narr2}
\alias{process_narr2}
\title{Process NARR Data (v2)}
\usage{
process_narr2(
  date = c("2023-09-01", "2023-09-01"),
  variable = NULL,
  path = NULL,
  ...
)
}
\arguments{
\item{date}{A character vector specifying the start and end dates
in the format "YYYY-MM-DD".}

\item{variable}{A character vector specifying the variable of interest.}

\item{path}{A character vector specifying the path to the data files.}

\item{...}{Additional parameters to be passed to other functions.}
}
\value{
A SpatRaster object containing the processed NARR2 data.
}
\description{
This function processes NARR2 data based on the specified parameters.
}
\details{
This function performs the following steps:
\enumerate{
\item Sets up the directory path.
\item Checks for null parameters.
\item Identifies file paths based on the specified variable.
\item Generates a date sequence based on the specified start and end dates.
\item Filters the file paths to include only dates of interest.
\item Sets up the search abbreviation and target variable.
\item Imports and processes the data for each file path.
\item Subsets the data to include only dates of interest.
\item Returns the processed data as a SpatRaster object.
}
}
\examples{
# Process NARR2 data for the variable "PRATE" from
# September 1, 2023 to September 1, 2023
\dontrun{
data <-
  process_narr2(
    date = c("2023-09-01", "2023-09-01"),
    variable = "PRATE",
    path = "/path/to/data"
  )
}
}
\keyword{Calculation}
