% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_postprocessing.R
\name{reduce_pca}
\alias{reduce_pca}
\title{Perform Principal Component Analysis}
\usage{
reduce_pca(data, threshold = 0.99)
}
\arguments{
\item{data}{data.frame or data.table}

\item{threshold}{numeric(1). A fraction of the total variance that should
be covered by the components.}
}
\value{
data.table with Principal Components sufficient to satisfy the
\code{threshold}.`
}
\description{
Perform Principal Component Analysis
}
\seealso{
\code{\link[recipes:step_pca]{recipes::step_pca()}}
}
\keyword{internal}
