% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_base_functions.R
\name{set_args_calc}
\alias{set_args_calc}
\title{Set arguments for the calculation process}
\usage{
set_args_calc(
  char_siteid = "site_id",
  char_timeid = "time",
  char_period = c("2018-01-01", "2022-10-31"),
  num_extent = c(-126, -62, 22, 52),
  char_user_email = paste0(Sys.getenv("USER"), "@nih.gov"),
  path_export = "inst/targets/punchcard_calc.qs",
  path_input = "input"
)
}
\arguments{
\item{char_siteid}{Character string specifying the site ID.
Default is "site_id".}

\item{char_timeid}{Character string specifying the time ID.
Default is "time".}

\item{char_period}{Character vector specifying the time period.
Default is c("2018-01-01", "2022-10-31").}

\item{path_export}{Character string specifying the export path.
Default is "inst/targets/punchcard_calc.qs".
If \code{NULL}, a list object "arglist_common" is exported to the global
environment and returns a list of arguments for the calculation process.}

\item{path_input}{Character string specifying the input path.
Default is "input".}

\item{extent}{Numeric vector specifying the extent.
Default is c(-126, -62, 22, 52).}

\item{user_email}{Character string specifying the user email.
Default is the current user's email.}
}
\value{
A list of arguments for the calculation process.
}
\description{
This function sets the arguments for the calculation process.
It takes several parameters including site ID, time ID, time period,
extent, user email, export path, and input path.
}
\note{
path_input should contain the following subdirectories:
\itemize{
\item modis/raw/61/MOD11A1
\item modis/raw/61/MOD06_L2
\item modis/raw/61/MOD09GA
\item modis/raw/61/MCD19A2
\item modis/raw/61/MOD13A2
\item modis/raw/5000/VNP46A2
\item aqs
\item geos/aqc_tavg_1hr_g1440x721_v1
\item geos/chm_tavg_1hr_g1440x721_v1
\item HMS_Smoke/data
\item gmted
\item nei
\item narr
\item HMS_Smoke
\item koppen_geiger
\item ecoregions
\item sedac_groads
\item sedac_population
}
}
\keyword{Utility}
