% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_learner.R
\name{switch_model}
\alias{switch_model}
\title{Define a base learner model based on parsnip and tune}
\usage{
switch_model(
  model_type = c("mlp", "xgb", "lgb", "elnet"),
  learn_rate = 0.1,
  device = "cuda:0"
)
}
\arguments{
\item{model_type}{character(1). Model type to be used.
Default is "mlp". Available options are "mlp", "xgb", "lgb", "elnet".}

\item{learn_rate}{numeric(1). The learning rate for the model.
Default is 0.1.}

\item{device}{character(1). The device to be used for training.
Default is "cuda:0". Make sure that your system is equipped
with CUDA-enabled graphical processing units.}
}
\value{
A parsnip model object.
}
\description{
Define a base learner model based on parsnip and tune
}
\keyword{Baselearner}
