% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_base_functions.R
\name{unmarshal_function}
\alias{unmarshal_function}
\title{Unmarshal functions}
\usage{
unmarshal_function(pkg_func_str)
}
\arguments{
\item{pkg_func_str}{Character string specifying the package and function.}
}
\value{
Function object.
}
\description{
Unmarshal functions
}
\examples{
unmarshal_function("amadeus::process_aqs")
}
\keyword{Utility}
