% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{unmarshal_function}
\alias{unmarshal_function}
\title{Unmarshal functions}
\usage{
unmarshal_function(pkg_func_str)
}
\arguments{
\item{pkg_func_str}{Character string specifying the package and function.}
}
\value{
Function object.
}
\description{
Unmarshal functions
}
\note{
The function name string must include two colons \code{::}.
Also, the package preceding the two colons should be loaded in the
current environment.
}
\examples{
unmarshal_function("amadeus::process_aqs")
}
\keyword{Utility}
