% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_base_functions.R
\name{vis_rset}
\alias{vis_rset}
\title{Visualize the spatio-temporal cross-validation index}
\usage{
vis_rset(rsplit, angle = 60)
}
\arguments{
\item{rsplit}{rsample::manual_rset() object.}

\item{angle}{numeric(1). Viewing angle of 3D plot.}
}
\value{
None. A plot will be generated.
}
\description{
Visualize the spatio-temporal cross-validation index
}
\seealso{
\code{\link{generate_cv_index}}
}
\keyword{Baselearner}
